% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effex-ggplot-s4-aesName.R
\docType{class}
\name{aesName-class}
\alias{aesName-class}
\alias{.aesName}
\alias{xAesName-class}
\alias{.xAesName}
\alias{yAesName-class}
\alias{.yAesName}
\alias{alphaAesName-class}
\alias{.alphaAesName}
\alias{colourAesName-class}
\alias{.colourAesName}
\alias{fillAesName-class}
\alias{.fillAesName}
\alias{groupAesName-class}
\alias{.groupAesName}
\alias{shapeAesName-class}
\alias{.shapeAesName}
\alias{sizeAesName-class}
\alias{.sizeAesName}
\alias{strokeAesName-class}
\alias{.strokeAesName}
\alias{aesName}
\title{aesName class}
\usage{
aesName(aes)
}
\arguments{
\item{aes}{The aes. If a subclass "<aes>AesName" exists, this will be
instantiated.}
}
\description{
In order to allow flexible inheritance, we define a dummy class for every
name of an aesthetic. Whereas any name may be used, the available methods
are:

\itemize{
\item x
}

\itemize{
\item y
}

\itemize{
\item alpha
}

\itemize{
\item colour
}

\itemize{
\item fill
}

\itemize{
\item group
}

\itemize{
\item shape
}

\itemize{
\item size
}

\itemize{
\item stroke
}
}
\details{
If other aesthetics are supplied, most methods default to a neutral
element, e. g. \code{identity} in the case of \code{\link[=fx_ggplot]{fx_ggplot()}}. You can build
\code{aesName} by calling the function and you can set its subclasses by
\code{new("<subclass>AesName", ...)}. This is, however, not recommended. It is
better to use the \code{sub} argument in \code{\link[=aesName]{aesName()}}, as it defaults to generating
\code{aesName} if the \code{sub} is unknown.
}
\examples{
aesName(aes = "x")
aesName(aes = "m")
aesName(aes = "colour")
}
