% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{architecture}
\alias{architecture}
\alias{as_architecture}
\title{Architecture}
\usage{
architecture(verbose = TRUE, ...)

as_architecture(archie, verbose = TRUE)
}
\arguments{
\item{verbose}{When the architecture's functions are called, should they
keep you up-to-date about what they are currently doing?}

\item{...}{Parameters of \code{new.env()}}

\item{archie}{an environment}
}
\description{
An architecture allows you to keep your variable system clean. Within the
architecture, you set global parameters of your indicator system.
}
\details{
Architectures allow a clear distinction between how you define your
variable system and what you need to load for a specific task. Just like the
call of \code{generate} in comparison to its functional definition, an
architecture is devoted to pure coding.
}
