% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-elements-dplyr_functions.R
\name{bind_coll_rows}
\alias{bind_coll_rows}
\title{Bind Variable Rows}
\usage{
bind_coll_rows(..., .id = NULL)
}
\arguments{
\item{.id}{Should a key be added that distinguishes between the different
data frames?}
}
\description{
\code{bind_coll_rows()} is a wrapper around \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}} that ensures
appropriate treatment of variables. Furthermore key variables retain their
class.
}
