% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-relation-class.R
\name{C_Relation}
\alias{C_Relation}
\title{C_Relation}
\usage{
C_Relation(relation, definition = "", ...)
}
\arguments{
\item{relation}{a character. the kind of relation}

\item{definition}{a short definition of the variable}

\item{...}{details on the relation}
}
\description{
C_Relations are special \code{content} variables that describe the relationship
between two \code{content}s (possible extension further down the road).
}
\seealso{
\code{\link[=Content]{Content()}}
}
