% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collection}
\alias{collection}
\alias{is_collection}
\alias{variables}
\alias{ind_names}
\alias{var_cols}
\alias{df_names}
\title{Collections}
\usage{
collection(...)

is_collection(x)

variables(x)

ind_names(x)

var_cols(x)

df_names(x)
}
\arguments{
\item{...}{objects of the class \code{collection_df}}

\item{x}{an object}
}
\description{
Collections are composed of collection elements (at the time only data
frames). Each collection can be uniquely identified by a key and/or a name,
where every unique combination of key and name provides a \code{collection_el}
with the respective key.
}
