% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-elements.R
\name{collection_df}
\alias{collection_df}
\alias{is_collection_df}
\title{Collection data frames}
\usage{
collection_df(df, key = NULL)

is_collection_df(x)
}
\arguments{
\item{df}{a data frame that will be transformed to a collection data frame.}

\item{key}{the key of the collection data frame. Needs to consist of uniquely
identifying columns of \code{df}. May also be \code{NULL} (default setting), in which
case a name needs to be supplied to the collection.}
}
\description{
Collection data frames are possible elements of a collection (represented by
the virtual class \code{collection_el}).
}
