% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-subsetting.R
\name{collection_subsetting}
\alias{collection_subsetting}
\alias{[.collection}
\title{Subsetting Collection}
\usage{
\method{[}{collection}(x, i)
}
\arguments{
\item{ind_names}{Select according to ind_name}

\item{cols}{Select according to ind_name}

\item{df_names}{Select according to df_name}

\item{df_keys}{Select according to df_key}

\item{nrs_lst}{This feature should be used with caution. It is mainly
available for internal purposes.}
}
\value{
A \code{collection}.
}
\description{
Collection subsetting essentially works on different columns of variables.
New functions may be written according to a standard template.
}
