% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{compare_keys}
\alias{compare_keys}
\alias{==.key}
\alias{!=.key}
\alias{<=.key}
\alias{<.key}
\alias{>=.key}
\alias{>.key}
\title{Compare keys}
\usage{
\method{==}{key}(e1, e2)

\method{!=}{key}(e1, e2)

\method{<=}{key}(e1, e2)

\method{<}{key}(e1, e2)

\method{>=}{key}(e1, e2)

\method{>}{key}(e1, e2)
}
\description{
Keys can be compared with all common comparison operators. In this context,
they are simply treated as sets. \code{==} returns whether they are equal and \code{<}
and \code{<=} return whether \code{e1} is an exclusive respectively inclusive subset of
\code{e2}.
}
