% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{complete_indicators}
\alias{complete_indicators}
\title{Complete indicators}
\usage{
complete_indicators(collection_df, indicators = indicator(), drop = FALSE)
}
\arguments{
\item{collection_df}{The \code{collection_df} that must be completed.}

\item{indicators}{The existing indicators.}

\item{drop}{logical. Should not applicable indicators be dropped.}
}
\description{
Complete indicators
}
