% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effex-ggplot.R
\name{fx_layer_complete_ind}
\alias{fx_layer_complete_ind}
\alias{def_fx_layer_complete_ind}
\title{Define layers dependent on all aesthetics}
\usage{
fx_layer_complete_ind(indicator, mapping, ind_names, ...)

def_fx_layer_complete_ind(fx_name, ..., x, y, alpha, colour, fill, group, shape,
  size, stroke)
}
\arguments{
\item{...}{other possible parameters. Not available for dispatch but can be
used in certain methods.}

\item{fx_name}{The \code{fxName} over which to dispatch}
}
\value{
A function that processes a \code{ggplot} object. Default is the identity
function.
}
\description{
These are the layers that depend on all aesthetics. If certain aesthetics are
not supplied or aesthetics of another class are given, this changes the
result. Due to the limitations of the S4 dispatch method, the aesthetics over
which it is possible to dispatch must be supplied in the generic. The
implemented methods are listed below:
}
