% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effex-ggplot.R,
%   R/effex-ggplot-layer-single_ind.R
\docType{methods}
\name{fx_layer_single_ind}
\alias{fx_layer_single_ind}
\alias{def_fx_layer_single_ind}
\alias{def_fx_layer_single_ind,aesName,geomFxName,fxDiscreteGeom-method}
\title{Define layers dependent on a single aesthetic}
\usage{
fx_layer_single_ind(indicator, aes_name, aes_quo, ind_names, ...)

def_fx_layer_single_ind(aes_name, fx_name, fx, indicator)

\S4method{def_fx_layer_single_ind}{aesName,geomFxName,fxDiscreteGeom}(aes_name,
  fx_name, fx)
}
\arguments{
\item{indicator}{The indicator information data frame.}

\item{aes_name}{The name of the aesthetic, previously processed by
\code{\link[=aesName]{aesName()}} to allow method dispatch.}

\item{fx_name}{The name of the effex element, previously processsed by
\code{\link[=fxName]{fxName()}} to allow method dispatch.}

\item{fx}{The corresponding fx element.}
}
\value{
A function that processes a \code{ggplot} object. Default is the identity
function.
}
\description{
These are the layers that are always present if a certain combination of
aesthetic and effex element is present. Below are listed all implemented
methods in \code{tectr}:
}
\details{
Method dispatch over both \code{fx_name} and \code{fx} might seem like
overkill, as they will often possess the identical signature. However, in
some cases, a subclass might allow for new processing methods.
}
