% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-elements-dplyr_functions.R
\name{dplyr_extensions}
\alias{dplyr_extensions}
\alias{select.collection_df}
\alias{mutate.collection_df}
\alias{arrange.collection_df}
\alias{filter.collection_df}
\alias{inner_join.collection_df}
\alias{left_join.collection_df}
\alias{right_join.collection_df}
\alias{full_join.collection_df}
\alias{semi_join.collection_df}
\alias{anti_join.collection_df}
\title{Extensions of dplyr methods}
\usage{
select.collection_df(.data, ...)

mutate.collection_df(.data, ...)

arrange.collection_df(.data, ...)

filter.collection_df(.data, ...)

inner_join.collection_df(x, y, by = NULL, copy = FALSE, suffix = c(".x",
  ".y"), ...)

left_join.collection_df(x, y, by = NULL, copy = FALSE, suffix = c(".x",
  ".y"), ...)

right_join.collection_df(x, y, by = NULL, copy = FALSE, suffix = c(".x",
  ".y"), ...)

full_join.collection_df(x, y, by = NULL, copy = FALSE, suffix = c(".x",
  ".y"), ...)

semi_join.collection_df(x, y, by = NULL, copy = FALSE, ...)

anti_join.collection_df(x, y, by = NULL, copy = FALSE, ...)
}
\description{
dplyr methods work differently on \code{collection_df}s, as they have to retain
key columns and ensure that the keys still uniquely determine observations.
For \code{collection_df}s, dplyr methods therefore never drop a key column,
preserve the right key in the case of namechanges and make sure that the
other conditions on \code{collection_df}s are satisfied. Details can be found in
the return section.
}
