% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effex-plot.R
\name{effex_plot}
\alias{effex_plot}
\title{Effex: Plot}
\usage{
effex_plot(data, indicator, metadata = NULL, mapping)
}
\arguments{
\item{metadata}{Defunct for now.}

\item{mapping}{Which dimensions do you wish to visualize by which indicators?}
}
\description{
This effex allows you to specify a mapping and the details get filled
in automatically by the indicator.
}
