% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effex-selectIndicator.R
\name{effex_selectIndicator}
\alias{effex_selectIndicator}
\title{Effex: Select Indicators}
\usage{
effex_selectIndicator(indicator, mapping, inputId)
}
\description{
This effex sets up a shiny element that allows you to select among different
indicators according to their name.
}
