% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-methods.R
\name{extend}
\alias{extend}
\alias{extend_variables}
\alias{extend_relations}
\alias{extend_edges}
\title{Extend a structure}
\usage{
extend(structure, variables = Content(character(0)),
  relations = C_Relation(character(0)), edges = data.frame(from =
  numeric(0), to = numeric(0), name = numeric(0)))

extend_variables(structure, variables = Content(character(0)))

extend_relations(structure, relations = C_Relation(character(0)))

extend_edges(structure, edges = data.frame(from = numeric(0), to = numeric(0),
  name = numeric(0)))
}
\arguments{
\item{structure}{a structure to be extended}

\item{variables}{The variables of the structure.}

\item{relations}{relations that extend the structure}

\item{edges}{new connection between contents and relations}
}
\description{
Extend a structure with new content, new relations or new edges.
}
