% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R
\name{filter_req_aes}
\alias{filter_req_aes}
\title{Filter according to the required aesthetics}
\usage{
filter_req_aes(nominations, mapping)
}
\description{
If the required aesthetics cannot be either found in the mapping or in the
mapping of the object itself, remove the layers nomination.
}
\details{
This function is superfluous for now, see \code{\link[=required_aes]{required_aes()}}.
}
