% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-default-fxGeom.R, R/fx-fun-default.R
\name{fxe_default.fxd_default_fxGeom_class}
\alias{fxe_default.fxd_default_fxGeom_class}
\alias{fxe_default.fxd_default_fxGeom_limits}
\alias{fxe_default.fxd_default_fxGeom_trans}
\alias{fx_default}
\alias{fxe_default}
\alias{fxe_default.fxd_default}
\title{Effex functions: Defaults for the metaframe}
\usage{
\method{fxe_default}{fxd_default_fxGeom_class}(data, mf, col, ...)

\method{fxe_default}{fxd_default_fxGeom_limits}(data, mf, col, ...)

\method{fxe_default}{fxd_default_fxGeom_trans}(data, mf, col, ...)

fx_default(data, columns = character(0), ..., overwrite = FALSE,
  mustWork = TRUE)

fxe_default(data, mf, col, ..., mustWork = TRUE)

\method{fxe_default}{fxd_default}(data, mf, col, ..., mustWork = TRUE)
}
\arguments{
\item{data}{The dataframe with a metaframe attribute}

\item{...}{Parameters to give on to the methods of \code{\link[=fxe_default]{fxe_default()}}}

\item{columns}{\code{character}. The columns of the metaframe that should be
provided.}

\item{overwrite}{\code{logical}. Should existing columns be overwritten? If \code{TRUE}
might also delete columns without compatible method.}

\item{mustWork}{If no compatible method is found, should the method return
\code{NULL} or throw an error?}
}
\description{
The \code{fx_default}-family consists of functions with a consistent interface that
that sets a default value for the specified metaframe column. Their intention
is an explicit standard mode
which allows to concentrate on changing special cases only.
\code{fx_default} itself wraps around these functions by modifying the metaframe
for any number of metaframe columns. This is, of course only useful if
modifications need not be applied.
}
\section{Methods (by generic)}{
\itemize{
\item \code{fxe_default}: infers the \code{fxGeom_class} via
\code{\link[=fx_default_fxGeom_class]{fx_default_fxGeom_class()}}

\item \code{fxe_default}: infers the column \code{fxGeom_limits} via
\code{\link[=fx_default_fxGeom_limits]{fx_default_fxGeom_limits()}}

\item \code{fxe_default}: infers the column \code{fxGeom_trans} which controls axis
transformations via \code{\link[=fx_default_fxGeom_trans]{fx_default_fxGeom_trans()}}.

\item \code{fxe_default}: returns an error if no previous method has been found.
}}

\section{Extension Mechanism}{

\code{fx_default} can be extended by adding a method to the S3 generic
\code{fxe_default}.
It is recommended to specify the particular method following the scheme
\code{fx_default_<column name>} to make it easier to use effex default outside an
entire data frame. The new method must conform to the \code{\link[=fxd]{fxd()}} naming
mechanism where the task is "default" and the column name is the subclass.
}

\examples{
fxe_default(data.frame(),
            new_metaframe(name = character(0)),
            col = fxd("default"),
            mustWork = FALSE)
\donttest{
fxe_default(data.frame(),
            new_metaframe(name = character(0)),
            col = fxd("default"))
}


}
