% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-default-fxGeom.R
\name{fx_default_fxGeom_class}
\alias{fx_default_fxGeom_class}
\title{Effex Functions: Default of the column \code{fxGeom_class}}
\usage{
fx_default_fxGeom_class(data, mf = metaframe(fx_default(data)),
  custom_fun = default_identity)
}
\arguments{
\item{data}{A data frame}

\item{mf}{The metaframe. The metaframe of \code{data} is set as default.}

\item{custom_fun}{A function which accepts the determined class and the name
and returns the corrected class. Default is the identity.}
}
\description{
\code{fxGeom_class} specifies the class of the \code{fxGeom} object which \code{\link[=fx_ggplot]{fx_ggplot()}}
works on. \code{fx_default_fxGeom_class} infers the class according to the class of
the corresponding column.
}
\section{Mechanism}{

\itemize{
\item Numeric: "Continuous"
\item Character or Factor: "Discrete"
\item Simple Feature column: "Spatial"
\item Else: "" (no class)
}
}

\examples{
df <- dplyr::tibble(
  Cont = 1:10,
  Disc = LETTERS[1:10],
  Disc2 = factor(LETTERS[1:10]),
  List = as.list(1:10)
)
fx_default_fxGeom_class(df)

}
