% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-default-fxGeom.R
\name{fx_default_fxGeom_limits}
\alias{fx_default_fxGeom_limits}
\title{Effex Functions: Default of \code{fxGeom_limits}}
\usage{
fx_default_fxGeom_limits(data, mf = metaframe(fx_default(data)),
  custom_fun = default_identity)
}
\description{
\code{fxGeom_limits} specifies the limits of the scale of \code{\link[=fx_ggplot]{fx_ggplot()}}.
}
\section{Mechanism}{

\itemize{
\item Numeric: The range of the data
\item Character: The unique values
\item Factor: The levels
\item Else: NULL
}
}

\examples{
df <- dplyr::tibble(
  num = c(1:3, 1:3),
  char = LETTERS[c(1:3, 1:3)],
  fac = factor(LETTERS[c(1:3, 1:3)], levels = LETTERS[1:4]),
  lst = as.list(c(1:3, 1:3))
)
fx_default_fxGeom_limits(df)

}
