% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-default-fxGeom.R
\docType{methods}
\name{fx_default_fxGeom_pal}
\alias{fx_default_fxGeom_pal}
\alias{fxe_default_fxGeom_pal}
\alias{fxe_default_fxGeom_pal,fxGeomContinuous,x_yAesName-method}
\alias{fxe_default_fxGeom_pal,fxGeomContinuous,colour_fillAesName-method}
\alias{fxe_default_fxGeom_pal,fxGeomContinuous,alphaAesName-method}
\alias{fxe_default_fxGeom_pal,fxGeomContinuous,sizeAesName-method}
\alias{fxe_default_fxGeom_pal,fxGeomDiscrete,x_yAesName-method}
\alias{fxe_default_fxGeom_pal,fxGeomDiscrete,colour_fillAesName-method}
\alias{fxe_default_fxGeom_pal,fxGeomDiscrete,shapeAesName-method}
\title{Effex Functions: Default of \code{fxGeom_pal}}
\usage{
fx_default_fxGeom_pal(data, mf = metaframe(fx_default(data)), aes_name)

fxe_default_fxGeom_pal(fx_geom, aes_name, ...)

\S4method{fxe_default_fxGeom_pal}{fxGeomContinuous,x_yAesName}(fx_geom,
  aes_name, ...)


  \S4method{fxe_default_fxGeom_pal}{fxGeomContinuous,colour_fillAesName}(fx_geom,
  aes_name, ...)

\S4method{fxe_default_fxGeom_pal}{fxGeomContinuous,alphaAesName}(fx_geom,
  aes_name, ...)

\S4method{fxe_default_fxGeom_pal}{fxGeomContinuous,sizeAesName}(fx_geom,
  aes_name, ...)

\S4method{fxe_default_fxGeom_pal}{fxGeomDiscrete,x_yAesName}(fx_geom,
  aes_name, ...)


  \S4method{fxe_default_fxGeom_pal}{fxGeomDiscrete,colour_fillAesName}(fx_geom,
  aes_name, ...)

\S4method{fxe_default_fxGeom_pal}{fxGeomDiscrete,shapeAesName}(fx_geom,
  aes_name, ...)

\S4method{fxe_default_fxGeom_pal}{fxGeomDiscrete,shapeAesName}(fx_geom,
  aes_name, ...)
}
\description{
The default palette depends on the aesthetic that is used and mostly
corresponds to the default choices in \link{ggplot2}. It is extendible via
\code{fxe_default_fxGeom_pal} over \code{fx_geom} and \code{aes_name}.
}
