% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-default-fxGeom.R
\name{fx_default_fxGeom_trans}
\alias{fx_default_fxGeom_trans}
\title{Effex Functions: Default of \code{fxGeom_trans}}
\usage{
fx_default_fxGeom_trans(data, mf = metaframe(fx_default(data)),
  fxGeom_trans_simple = TRUE, fxGeom_trans_p.threshold = 0.01)
}
\arguments{
\item{fxGeom_trans_simple}{\code{logical} should all boxcox and modulus
transformations be allowed or should they be restricted to "identity",
"sqrt" and "log10"?}

\item{fxGeom_trans_p.threshold}{a value between 0 and 1. How low does the
p-value of the \code{\link[moments:agostino.test]{moments::agostino.test()}} have to be to make another
transformation necessary? A rather low value is recommended.}
}
\description{
The default transformation is determined in a manner that is sensitive to
skewness. Skewed distribution are often hard to interpret if the
axes are not transformed.
}
\section{Mechanism}{

This function has two different modi: \code{fxGeom_trans_simple = TRUE/FALSE}.
\code{TRUE} means that the only possible transformations are "identity", "sqrt"
and "log10". \code{FALSE}, on the other hand, allows for the entire range of
boxcox transformations.

Non-numeric data is always assigned "identity".

If \code{fxGeom_trans_simple} is true, the skewness of the three transformation is
compared and the lowest absolute skewness wins. If there is nonpositive
numeric data, "identity" is chosen by default.

If \code{fxGeom_trans_simple} is false, the procedure consists of severals steps:
\itemize{
\item To avoid unnecessary transformations, the \code{\link[moments:agostino.test]{moments::agostino.test()}} for
unnormal skewness is conducted. If the corresponding p-value is not below
\code{fxGeom_trans_p.threshold}, the transformation is "identity".
\item Afterwards, \code{\link[geoR:boxcoxfit]{geoR::boxcoxfit()}} determines a suitable transformation
which is then transformed into a suitable object by \code{\link[scales:boxcox_trans]{scales::boxcox_trans()}}
}
}

\examples{
set.seed(1)
norm <- rnorm(1000)
square <- norm^2
exp <- exp(norm)
df <- dplyr::tibble(
  norm = rnorm(100),
  square = rnorm(100)^2,
  exp = exp(rnorm(100)),
  chars = rep("a", 100)
)
fx_default_fxGeom_trans(df)
fx_default_fxGeom_trans(fxGeom_trans_simple = FALSE)

}
\seealso{
\code{\link[scales:boxcox_trans]{scales::boxcox_trans()}}
}
