% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-default-fxInfo.R
\name{fx_default_fxInfo_name}
\alias{fx_default_fxInfo_name}
\title{Effex Functions: Default for the column \code{fxInfo_name}}
\usage{
fx_default_fxInfo_name(data, mf)
}
\arguments{
\item{data}{a data frame}

\item{mf}{a metaframe}
}
\description{
\code{fxInfo_name} is a more human-readable title of the particular variable than
\code{name}. Nevertheless, inference simply assigns the name to it.
}
