% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-fx_mtcars.R
\docType{data}
\name{fx_diamonds}
\alias{fx_diamonds}
\title{diamonds indicators}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 5 columns.}
\usage{
fx_diamonds
}
\description{
This tibble provides an example for an indicator for the \code{diamonds} dataset.
}
\keyword{datasets}
