% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_factor.R
\name{fx_evaluate}
\alias{fx_evaluate}
\title{Evaluating factors}
\usage{
fx_evaluate(x)
}
\description{
\code{fx_evaluate} returns the list/vector of the original kind. It is defined as
a generic to make uncomplicated evaluation of a data frame possible. \code{[}
provides ordinary subsetting functions whereas \code{[[} automatically evaluates.
}
