% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_factor.R
\name{fx_factor}
\alias{fx_factor}
\alias{as_fx_factor}
\alias{is_fx_factor}
\alias{levels<-.fx_factor}
\alias{fx_evaluate.default}
\alias{fx_evaluate.fx_factor}
\alias{fx_evaluate.data.frame}
\alias{fx_evaluate.metaframe}
\alias{[.fx_factor}
\alias{[[.fx_factor}
\title{fx_factor: A simple, downsized alternative for arbitrary objects}
\usage{
fx_factor(x, levels = unique(x))

as_fx_factor(x_int, levels)

is_fx_factor(x)

\method{levels}{fx_factor}(x) <- value

\method{fx_evaluate}{default}(x)

\method{fx_evaluate}{fx_factor}(x)

\method{fx_evaluate}{data.frame}(x)

\method{fx_evaluate}{metaframe}(x)

\method{[}{fx_factor}(x, i)

\method{[[}{fx_factor}(x, i)
}
\arguments{
\item{x}{a vector or a list}

\item{levels}{a set of possible, unique objects that may occur in x}

\item{x_int}{an integer vector referring to the levels}
}
\description{
Often, objects of a large size repeating themselves need to be stored in a
data frame. \code{fx_factor} provides a simple framework to work with these
objects in a space-saving manner - at least until the modeling and
visualization.
}
\details{
Levels inference is more complicated for arbitrary classes. Requiring
\href{identity}{identical()} may be too much (e. g.
\code{identical(ggplot2::geom_point(), ggplot2::geom_point())}) whereas \code{\link[=match]{match()}}
(matching) might be too lax (e. g.
\code{match(list(ggplot2::geom_point()), list(ggplot2::geom_point(stat = "log10"))}).
If in doubt, using \code{as_fx_factor} is recommended and \code{fx_factor} will return
an error if encoding and decoding changes the result according to \code{identical}
(with the argument \code{ignore.environment = TRUE}).

Also, all elements of \code{x} must occur in \code{levels}.
}
\examples{


}
