% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effex-ggplot.R
\name{fx_ggplot}
\alias{fx_ggplot}
\title{Effex: ggplot}
\usage{
fx_ggplot(data, indicator, aes_mapping, ind_names, geom, ...)
}
\arguments{
\item{data}{The data that is visualized.}

\item{indicator}{The indicators that specify the effex.}

\item{aes_mapping}{A mapping of \href{aesthetics}{effex_aes()} to indicators. This
should work in the same way as the selection mechanism that can for instance
be found in \code{\link[dplyr:select]{dplyr::select()}}. \code{effex_aes(x = starts_with("v2"))} would
therefore be valid.}

\item{ind_names}{the effex that defines the names of the indicators}

\item{...}{other non-crucial (but possibly important, e. g. description)
effex.}

\item{geoms}{the effex that defines the geoms.}
}
\description{
This effex lets you build visualization plots that depend on the specified
aesthetics dynamically.
}
