% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-infer.R
\name{fx_infer}
\alias{fx_infer}
\alias{fxint_infer}
\alias{fxext_infer}
\alias{fxext_infer.fxd_infer}
\title{Effex functions: Infer the metaframe}
\usage{
fx_infer(data, columns = character(0), ..., overwrite = FALSE,
  mustWork = TRUE)

fxint_infer(data, col, ...)

fxext_infer(data, mf, col, ..., mustWork = TRUE)

\method{fxext_infer}{fxd_infer}(data, mf, col, ..., mustWork = TRUE)
}
\arguments{
\item{data}{The dataframe with a metaframe attribute}

\item{columns}{\code{character}. The columns of the metaframe that should be
inferred.}

\item{...}{Parameters to give on to the methods of \code{\link[=fxext_infer]{fxext_infer()}}}

\item{overwrite}{\code{logical}. Should existing columns be overwritten? If \code{TRUE}
might also delete columns without compatible method.}

\item{mustWork}{If no compatible method is found, should the method return
\code{NULL} or throw an error?}
}
\description{
The \code{fx_infer}-family consists of functions with a consistent interface that
attempts automatical inference for the specified metaframe column. They
return the inferred column. Their intention is an explicit standard mode
which allows to concentrate on changing special cases only.
\code{fx_infer} itself wraps around these functions by modifying the metaframe
for any number of metaframe columns. This is, of course only useful if
modifications need not be applied.
}
\details{
\code{fx_infer} can be extended by adding a method to the generic \code{fxext_infer}.
It is recommended to specify the particular method following the scheme
\code{fx_infer_<column name>} to make it easier to use effex inference outside an
entire data frame. To achieve that \code{fxint_infer} creates a structure of class
<column name> with superclass "fx_infer" via \code{\link[=fxd]{fxd()}}
}
