% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-infer-fxGeom.R
\name{fx_infer_fxGeom_class}
\alias{fx_infer_fxGeom_class}
\title{Effex Functions: Infer the column \code{fxGeom_class}}
\usage{
fx_infer_fxGeom_class(data, mf)
}
\arguments{
\item{data}{a data frame}

\item{mf}{a metaframe}
}
\description{
\code{fxGeom_class} specifies the class of the \code{fxGeom} object which \code{\link[=fx_ggplot]{fx_ggplot()}}
works on. \code{fx_infer_fxGeom_class} infers the class according to the class of
the corresponding column.
}
\section{Inference Mechanism}{

\itemize{
\item Numeric: \href{fxGeomContinuous}{fxGeomContinuous-class}
\item Character or Factor: \href{fxGeomDiscrete}{fxGeomDiscrete-class}
\item Simple Feature column: \href{fxGeomSpatial}{fxGeomSpatial-class}
\item Else: \href{fxGeom}{fxGeom-class}
}
}

