% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-info.R
\name{fx_info}
\alias{fx_info}
\title{Effex Function: Information}
\usage{
fx_info(data, topic, ...)
}
\arguments{
\item{data}{A dataframe, possibly with a metaframe}

\item{topic}{The topic of the information}

\item{arguments}{for the information methods}
}
\description{
This function provides versatile information on the different variables.
\code{\link[=fx_output]{fx_output()}} turns this information into a string.
}
