% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-info-title.R
\name{fx_info_title}
\alias{fx_info_title}
\title{Build up a variable title}
\usage{
fx_info_title(data, name, ..., fxInfo_title = NULL,
  fxInfo_title_fun = NULL, fxInfo_name, fxInfo_unit = NULL,
  fxInfo_title_stats = character(0), fxInfo_title_na.show = NULL,
  fxInfo_title_na.show.threshold = 1L, fxInfo_title_n.show = FALSE,
  fxInfo_title_unit.show = FALSE)
}
\arguments{
\item{data}{a data frame, possibly with metaframe}

\item{name}{the internal name of the variable}

\item{fxInfo_title}{either a title or \code{NULL}. If it is not \code{NULL}, all other
parameters are irrelevant.}

\item{fxInfo_title_fun}{either a function which determines a title or \code{NULL}.
If it is not \code{NULL}, all parameters below are irrelevant.}

\item{fxInfo_name}{the name of the variable}

\item{fxInfo_unit}{the unit of the variable or \code{NULL}}

\item{fxInfo_title_stats}{the statistics that should be given in the title.
Number of data points and missing values are given separately.}

\item{fxInfo_title_na.show}{Should the number of missing values be displayed?}

\item{fxInfo_title_na.show.threshold}{If \code{fxInfo_title_na.show} is missing,
what is the
threshold number of missing values to display them. Default is 1.}

\item{fxInfo_title_n.show}{Should the number of data points be displayed.
Default is \code{FALSE}}

\item{fxInfo_title_unit.show}{Should the unit be displayed?}
}
\description{
This function builds a title from a name and possibly certain statistics and
units.
}
