% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-output.R
\name{fx_output}
\alias{fx_output}
\alias{fxe_output}
\alias{fxe_output.fxd_output_}
\alias{fxe_output.fxd_output_asis}
\alias{fxe_output.fxd_output_table}
\alias{fxe_output.fxd_output_collapse}
\title{Effex Function: Render Output}
\usage{
fx_output(info, form = "", out_format = "rst", ...)

fxe_output(info, form, out_format, ...)

\method{fxe_output}{fxd_output_}(info, form, out_format, ...)

\method{fxe_output}{fxd_output_asis}(info, form, out_format, ...)

\method{fxe_output}{fxd_output_table}(info, form, out_format, out_file,
  digits = 2, ...)

\method{fxe_output}{fxd_output_collapse}(info, form, out_format, out_file,
  cell_scheme = if (out_format == "latex") "<<name>>: <<value>>" else
  "{name}: {value}", digits = 2, cell_sep = ", ", .open = if
  (out_format == "latex") "<<" else "{", .close = if (out_format ==
  "latex") ">>" else "}", ...)
}
\arguments{
\item{info}{The info tibble}

\item{form}{What form should it take. \code{fx_output} can be extended via this
argument (see \code{\link[=fxe_output]{fxe_output()}})}

\item{out_format}{What format should the output be in? Default argument is
"rst" which means an R object. For now, the other valid values are
"markdown", "html" and "latex".}

\item{...}{arguments for the \code{\link[=fxe_output]{fxe_output()}} methods}

\item{out_file}{Should the object be saved somewhere? If yes, provide the
path - the object will still be returned in the end.}

\item{cell_scheme}{a \code{character} that specifies how every cell, in connection
with its column name is styled. \code{name} enclosed by the \code{character}s specified
in \code{.open} and \code{.close} (see the documentation of \code{\link[glue:glue]{glue::glue()}}) refers
to the column name and \code{value} to the cell content. Note
that you will have to specify \code{cell_scheme} if you change \code{.open} or
\code{.close}.}

\item{cell_spec}{}
}
\description{
This function transforms the tibble which \code{\link[=fx_info]{fx_info()}} yields (and which may
be modified afterwards) into a certain form.
}
\section{Forms}{

\itemize{
\item \code{form = ""}: This standard does nothing except removing the name column
which should generally only be used for internal purposes. It is only
compatible with \code{out_format = NULL}.
}


\itemize{
\item \code{form = "asis"}: This form changes absolutely nothing. It only works with
\code{out_format = NULL}, as well.
}


\itemize{
\item \code{form = "table"}: This form generates a table and
works with all output formats. It is mainly powered by \code{\link[knitr:kable]{knitr::kable()}} and
you can pass additional options to both \code{kable} and
\code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}} via \code{...}
}


\itemize{
\item \code{form = "collapse"}: This form generates one line for every row of the
tibble. \code{...} are passed on to \code{\link[=format]{format()}}.
}
}

