% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-read.R
\name{fx_read}
\alias{fx_read}
\title{Effex functions: Read Data}
\usage{
fx_read(data)
}
\description{
The counterpart to the \link{fx_write}-functions. Recreates the original data
frame from any filepath columns.
}
