% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxd-obj.R
\name{fxd}
\alias{fxd}
\alias{is_fxd}
\alias{fxd_task}
\alias{fxd_subclass}
\title{Effex Dispatch}
\usage{
fxd(task, subclass = NULL)

is_fxd(x)

fxd_task(x)

fxd_subclass(x)
}
\arguments{
\item{task}{a character without "_" that can specify a class}

\item{subclass}{a character that can specify a class (may include "_").
Default is NULL which indicates that a subclass is unnecessary.}

\item{x}{an object}
}
\description{
Effex Dispatch objects are intended for a simplified dispatch with S3
generics. They have the classes "fxd_<task>\emph{<subclass>" where task is a
specific task, often fulfilled by the function "fx}<task>" (e. g. \code{\link[=fx_info]{fx_info()}})
and subclass stands for a specific category that is useful within the task.
}
