% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-info-statistics.R
\name{fxe_info.fxd_info_stats}
\alias{fxe_info.fxd_info_stats}
\title{Effex Function: Information -- Statistical Description}
\usage{
\method{fxe_info}{fxd_info_stats}(data, topic, statistics, ...)
}
\arguments{
\item{statistics}{a character vector or a list of characters and functions.
Function slots must have a name. Functions can accept arguments from
metaframe columns. The corresponding column will be fed in first and there
will be an additional data column with the entire data. All arguments but
the column will only be fed into functions, not characters.}
}
\description{
Effex Function: Information -- Statistical Description
}
