% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-info-title.R
\name{fxe_info.fxd_info_title}
\alias{fxe_info.fxd_info_title}
\alias{fx_info_title}
\title{Effex Function: Information -- Titles}
\usage{
\method{fxe_info}{fxd_info_title}(data, topic, ...)

fx_info_title(data, name, ..., fxInfo_title = NULL,
  fxInfo_title_fun = NULL, fxInfo_name, fxInfo_unit = NULL,
  fxInfo_title_stats = character(0), fxInfo_title_na.show = NULL,
  fxInfo_title_na.show.threshold = 1L, fxInfo_title_n.show = FALSE,
  fxInfo_title_unit.show = FALSE)
}
\arguments{
\item{fxInfo_title}{\code{character}, \code{NULL} or list of both. If}
}
\description{
These functions provide different kinds of titles
}
