% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-info.R
\name{fxe_info}
\alias{fxe_info}
\alias{fxe_info.fxd_info}
\title{Effex Externals: Information}
\usage{
fxe_info(data, topic, ...)

\method{fxe_info}{fxd_info}(data, topic, ...)
}
\arguments{
\item{data}{A dataframe, possibly with a metaframe}

\item{topic}{The topic of the information}
}
\description{
This generic powers \code{\link[=fx_info]{fx_info()}}. A new topic can be supplied by providing a
method for the class "fxd_info_<topic>".
}
