% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R
\docType{methods}
\name{fxe_labeller}
\alias{fxe_labeller}
\alias{fxe_labeller,fxGeomDiscrete-method}
\title{Effex Extendibles: \code{fx_ggplot} facetting: labellers}
\usage{
fxe_labeller(fxGeom_class, name, ...)

\S4method{fxe_labeller}{fxGeomDiscrete}(fxGeom_class, name, fxGeom_limits,
  fxGeom_breaks = NULL, ...)
}
\value{
a valid argument to \code{\link[ggplot2:labeller]{ggplot2::labeller()}}. Supplies \code{\link[=fx_ggplot2]{fx_ggplot2()}}
with the labels for the facetting variable
}
\description{
Effex Extendibles: \code{fx_ggplot} facetting: labellers
}
