% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R,
%   R/fx-fun-ggplot-fxGeomContinuous.R, R/fx-fun-ggplot-fxGeomDiscrete.R,
%   R/fx-fun-ggplot-fxGeomSpatial.R, R/fx-fun-ggplot-fxGeomTime.R
\docType{methods}
\name{fxe_layer_complete_nominate}
\alias{fxe_layer_complete_nominate}
\alias{fxe_layer_complete_nominate,fxGeom,AesName-method}
\alias{fxe_layer_complete_veto}
\alias{fxe_layer_complete_vote}
\alias{fxe_layer_complete_nominate,fxGeomContinuous,xAesName-method}
\alias{fxe_layer_complete_nominate,fxGeomContinuous,yAesName-method}
\alias{fxe_layer_complete_nominate,fxGeomDiscrete,xAesName-method}
\alias{fxe_layer_complete_nominate,fxGeomDiscrete,yAesName-method}
\alias{fxe_layer_complete_nominate,fxGeomSpatial,geomAesName-method}
\alias{fxe_layer_complete_nominate,fxGeomTime,xAesName-method}
\title{Effex Extendibles: \code{fx_ggplot} - nominate layers}
\usage{
fxe_layer_complete_nominate(fx_geom, aes_name, data, ...)

\S4method{fxe_layer_complete_nominate}{fxGeom,AesName}(fx_geom, aes_name,
  data, ..., fxGeom_nominations = NULL)

fxe_layer_complete_veto(nomination, fx_geom, aes_name, data, ...)

fxe_layer_complete_vote(nomination, fx_geom, aes_name, data, ...)

\S4method{fxe_layer_complete_nominate}{fxGeomContinuous,xAesName}(fx_geom,
  aes_name, data, ..., name, fxGeom_nominations = NULL,
  fxGeom_alpha.threshold = NULL, fxGeom_alpha.half = NULL,
  fxGeom_alpha.min = NULL)

\S4method{fxe_layer_complete_nominate}{fxGeomContinuous,yAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL,
  fxGeom_alpha.threshold = NULL, fxGeom_alpha.half = NULL,
  fxGeom_alpha.min = NULL)

\S4method{fxe_layer_complete_nominate}{fxGeomDiscrete,xAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL)

\S4method{fxe_layer_complete_nominate}{fxGeomDiscrete,yAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL)

\S4method{fxe_layer_complete_nominate}{fxGeomSpatial,geomAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL)

\S4method{fxe_layer_complete_nominate}{fxGeomTime,xAesName}(fx_geom,
  aes_name, data, ..., name, fxGeom_nominations = NULL,
  fxGeom_alpha.threshold = NULL, fxGeom_alpha.half = NULL,
  fxGeom_alpha.min = NULL, fxGeom_assoc_vars = NULL,
  fxGeom_shadow.threshold = NULL, fxGeom_shadow.alpha = NULL)
}
\arguments{
\item{nomination}{a \code{\link[=nomination]{nomination()}}}

\item{fxGeom_shadow.threshold}{What is the threshold for the number of groups
so that the different lines are greyed out and a smoother is fitted?}
}
\description{
These functions drive \code{\link[=fxi_layer_complete]{fxi_layer_complete()}}. Provided that they fulfill
certain conditions they can be extended.
}
\details{
\code{fxe_layer_complete_nominate} returns a list of
\code{\link[=nomination]{nomination()}}(layer nominations). The default returns either an empty
list or, if provided the element of the metaframe column
\code{fxGeom_nominations}.

\code{fxe_layer_complete_veto} returns a logical value that indicates whether
this layer should be vetoed, i. e. removed from the nominations list. The
default method returns either \code{FALSE} or, if supplied with an element of the
metaframe column, it applies that function.
}
\seealso{
\code{\link[=fxi_layer_complete]{fxi_layer_complete()}}
}
