% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R,
%   R/fx-fun-ggplot-fxGeomContinuous.R, R/fx-fun-ggplot-fxGeomContinuousCI.R,
%   R/fx-fun-ggplot-fxGeomDiscrete.R, R/fx-fun-ggplot-fxGeomOrdinalCI.R,
%   R/fx-fun-ggplot-fxGeomTime.R
\docType{methods}
\name{fxe_layer_complete_vote}
\alias{fxe_layer_complete_vote}
\alias{fxe_layer_complete_vote,ANY,fxGeom,AesName-method}
\alias{fxe_layer_complete_vote,ANY,fxGeomContinuous,xAesName-method}
\alias{fxe_layer_complete_vote,ANY,fxGeomContinuous,yAesName-method}
\alias{fxe_layer_complete_vote,ANY,fxGeomContinuousCI,yAesName-method}
\alias{fxe_layer_complete_vote,ANY,fxGeomDiscrete,xAesName-method}
\alias{fxe_layer_complete_vote,ANY,fxGeomDiscrete,yAesName-method}
\alias{fxe_layer_complete_vote,ANY,fxGeomOrdinalCI,yAesName-method}
\alias{fxe_layer_complete_vote,ANY,fxGeomTime,xAesName-method}
\title{Effex Extendibles in \code{fx_ggplot}: Vote for the nominations}
\usage{
fxe_layer_complete_vote(nomination, fx_geom, aes_name, data, ...)

\S4method{fxe_layer_complete_vote}{ANY,fxGeom,AesName}(nomination, fx_geom,
  aes_name, data, ..., fxGeom_vote = NULL)


  \S4method{fxe_layer_complete_vote}{ANY,fxGeomContinuous,xAesName}(nomination,
  fx_geom, aes_name, data, ..., fxGeom_vetos = NULL,
  fxGeom_votes = NULL, fxGeom_alpha.threshold = NULL,
  fxGeom_alpha.half = NULL, fxGeom_alpha.min = NULL,
  fxGeom_hex.threshold = NULL)


  \S4method{fxe_layer_complete_vote}{ANY,fxGeomContinuous,yAesName}(nomination,
  fx_geom, aes_name, data, ..., fxGeom_vetos = NULL,
  fxGeom_votes = NULL, fxGeom_alpha.threshold = NULL,
  fxGeom_alpha.half = NULL, fxGeom_alpha.min = NULL,
  fxGeom_hex.threshold = NULL)


  \S4method{fxe_layer_complete_vote}{ANY,fxGeomContinuousCI,yAesName}(nomination,
  fx_geom, aes_name, data, ..., fxGeom_vetos = NULL,
  fxGeom_votes = NULL, fxGeom_alpha.threshold = NULL,
  fxGeom_alpha.half = NULL, fxGeom_alpha.min = NULL,
  fxGeom_hex.threshold = NULL)


  \S4method{fxe_layer_complete_vote}{ANY,fxGeomDiscrete,xAesName}(nomination,
  fx_geom, aes_name, data, ..., fxGeom_vetos = NULL,
  fxGeom_votes = NULL)


  \S4method{fxe_layer_complete_vote}{ANY,fxGeomDiscrete,yAesName}(nomination,
  fx_geom, aes_name, data, ..., fxGeom_vetos = NULL,
  fxGeom_votes = NULL)


  \S4method{fxe_layer_complete_vote}{ANY,fxGeomOrdinalCI,yAesName}(nomination,
  fx_geom, aes_name, data, ..., fxGeom_vetos = NULL,
  fxGeom_votes = NULL, fxGeom_alpha.threshold = NULL,
  fxGeom_alpha.half = NULL, fxGeom_alpha.min = NULL,
  fxGeom_hex.threshold = NULL)

\S4method{fxe_layer_complete_vote}{ANY,fxGeomTime,xAesName}(nomination,
  fx_geom, aes_name, data, ..., fxGeom_vetos = NULL,
  fxGeom_votes = NULL, fxGeom_alpha.threshold = NULL,
  fxGeom_alpha.half = NULL, fxGeom_alpha.min = NULL,
  fxGeom_hex.threshold = NULL)
}
\arguments{
\item{nomination}{the \code{\link[=nomination]{nomination()}} which is to be considered for a veto}

\item{fx_geom}{An S4 class argument which inherits from \linkS4class{fxGeom}.}

\item{aes_name}{An S4 class argument which inherits from \linkS4class{AesName}.}

\item{fxGeom_vote}{Either \code{NULL} (the default) or a function which returns
the votes as a numerical value for a given nomination.}

\item{fxGeom_alpha.threshold}{What is the threshold number of data points to
let transparency increase? Default is 1000}

\item{fxGeom_alpha.half}{What number of data points should result in
\code{alpha = 0.5}? Default is 3000}

\item{fxGeom_alpha.min}{What is the minimal \code{alpha}? Default is 0.05.}

\item{fxGeom_hex.threshold}{What is the threshold number of data points for
a hexagonal heatmap?}
}
\value{
A number which represents the amount of votes for a certain
nomination.
}
\description{
This function gathers the votes for the different nomination.
}
\section{Methods (by class)}{
\itemize{
\item \code{nomination = ANY,fx_geom = fxGeom,aes_name = AesName}: applies supplied functions to determine
additional votes

\item \code{nomination = ANY,fx_geom = fxGeomContinuous,aes_name = xAesName}: \itemize{
\item \code{GeomPoint} with the correct transparency (see \code{\link[=get_alpha]{get_alpha()}}): 1
\itemize{
\item hexagonal heatmap if the threshold is exceeded: 2 (else: 0)
\item histogram: 2
\item \code{StatBin} without \code{GeomBar}: 1
}
}

\item \code{nomination = ANY,fx_geom = fxGeomContinuous,aes_name = yAesName}: \itemize{
\item \code{GeomPoint} with the correct transparency (see \code{\link[=get_alpha]{get_alpha()}}): 1
\itemize{
\item hexagonal heatmap if the threshold is exceeded: 2 (else: 0)
\item \code{StatBin}: 1
\item boxplot: 2
}
}

\item \code{nomination = ANY,fx_geom = fxGeomContinuousCI,aes_name = yAesName}: \itemize{
\item line plot with shaded confidence intervals: 1
\itemize{
\item 
}
}

\item \code{nomination = ANY,fx_geom = fxGeomDiscrete,aes_name = xAesName}: \itemize{
\item heatmap: 1
\itemize{
\item barplot: 2
\item boxplot: 1
}
}

\item \code{nomination = ANY,fx_geom = fxGeomDiscrete,aes_name = yAesName}: \itemize{
\item heatmap: 1
}

\item \code{nomination = ANY,fx_geom = fxGeomOrdinalCI,aes_name = yAesName}: \itemize{
\item 
\itemize{
\item 
}
}

\item \code{nomination = ANY,fx_geom = fxGeomTime,aes_name = xAesName}: \itemize{
\item a line plot: 4
\itemize{
\item a path plot which is not a line plot: 3
}
}
}}

