% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-single-other.R
\docType{methods}
\name{fxe_layer_other}
\alias{fxe_layer_other}
\alias{fxe_layer_other,fxGeom,AesName-method}
\title{Effex Externals: Other single layer arguments}
\usage{
fxe_layer_other(fx_geom, aes_name, ...)

\S4method{fxe_layer_other}{fxGeom,AesName}(fx_geom, aes_name, ...)
}
\description{
This external function is used to add single layer arguments that are not
related to the scale. Note that there is no strict restriction for actual
\code{Scale}-arguments in \code{\link[=fxe_layer_scale]{fxe_layer_scale()}} and e. g. slight theme tweaks
might make sense sometimes.
}
