% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot.R
\name{fxext_layer_complete_nominate}
\alias{fxext_layer_complete_nominate}
\alias{fxext_layer_complete_vote}
\title{Effex Extendibles: \code{fx_ggplot} - nominate layers}
\usage{
fxext_layer_complete_nominate(fx_geom, aes_name, data, ...)

fxext_layer_complete_vote(geoms, fx_geom, aes_name, data, ...)
}
\arguments{
\item{geoms}{a data.frame with the columns \code{geom}, \code{stat} and \code{position} which
contain the nominations and \code{votes} which contain the votes that have been
given so far.}
}
\description{
\code{fxint_layer_complete} is mainly driven by two extendible functions:
\code{\link[=fxext_layer_complete_nominate]{fxext_layer_complete_nominate()}} which allows you to supply a data frame
with new nominations in the corresponding column for geom, stat, position and
params, and \code{\link[=fxext_layer_complete_vote]{fxext_layer_complete_vote()}} which allows you to give votes
depending on the geom, stat, position and params by modifying the \code{votes}
columns so that it adds your votes. If a certain combination is incompatible
with the data type, you may remove it.
}
\seealso{
fxint_layer_complete
}
