% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-scale.R
\docType{methods}
\name{fxext_layer_scale}
\alias{fxext_layer_scale}
\alias{fxext_layer_scale,fxGeom,AesName-method}
\alias{fxext_layer_scale,fxGeomContinuous,AesName-method}
\title{Effex External: Scale layers}
\usage{
fxext_layer_scale(fx_geom, aes_name, ...)

\S4method{fxext_layer_scale}{fxGeom,AesName}(fx_geom, aes_name, data,
  fxGeom_title = NULL, fxGeom_limits = c(NA_real_, NA_real_), ...)

\S4method{fxext_layer_scale}{fxGeomContinuous,AesName}(fx_geom, aes_name,
  fxGeom_breaks = NULL, fxGeom_minor_breaks = NULL,
  fxGeom_labels = NULL, fxGeom_expand = NULL, fxGeom_na.value = NULL,
  fxGeom_trans = "identity", ...)
}
\arguments{
\item{fx_geom}{the \code{\link[=fxGeom]{fxGeom()}} object}

\item{aes_name}{the \code{\link[=aesName]{aesName()}} object}

\item{fxGeom_title}{\code{NULL} or a character that provides the title}
}
\description{
This function provides an extendible scheme for scales for a single layer.
If some features are to be added, it is often sensible to \code{callNextMethod()}
and then add these features.
}
\section{Methods}{

Implemented methods are:


\itemize{
\item \code{fx_geom = fxGeom(), aes_name = AesName()}: Adds labels and limits
}


\itemize{
\item \code{fx_geom = fxGeom("Continuous"), aes_name = AesName("")}:
}
}

