% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot.R
\name{fxext_layer_single}
\alias{fxext_layer_single}
\title{Effex Extendibles: \code{fx-ggplot} - geometric information along one dimension}
\usage{
fxext_layer_single(fx_geom, aes_name, data, ...)
}
\description{
This is the extendible function for the single layer. It is dispatched
along the arguments \code{fx_geom} and \code{aes_name}. Arbitrary other arguments may
be added that should then be expected in the metaframe. It should always be
considered a possibility that these values are missing.
}
