% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R
\name{fxi_labeller}
\alias{fxi_labeller}
\title{Effex Internals: \code{fx_ggplot} facetting: labellers}
\usage{
fxi_labeller(data, facet_vars)
}
\description{
Effex Internals: \code{fx_ggplot} facetting: labellers
}
