% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R
\name{fxi_layer_single}
\alias{fxi_layer_single}
\title{Effex Internals: \code{fx-ggplot} - geometric information along one dimension}
\usage{
fxi_layer_single(data, mapping, nam)
}
\description{
Effex Internals: \code{fx-ggplot} - geometric information along one dimension
}
