% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxIndName.R
\docType{class}
\name{fxIndName-class}
\alias{fxIndName-class}
\alias{.fxIndName}
\alias{fxIndName}
\title{Effex Class: Indicator Names}
\usage{
fxIndName(ind_name)
}
\description{
This effex class only contains a single slot: \code{ind_name} which is defined by
a quosure that defines where to find the ind_names.
}
