% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R
\name{fxint_layer_complete}
\alias{fxint_layer_complete}
\alias{fxint_layer_complete_nominate}
\alias{fxint_layer_complete_vote}
\title{Effex Internals: \code{fx_ggplot} - combine all different dimensions}
\usage{
fxint_layer_complete(data, mapping)

fxint_layer_complete_nominate(data, mf)

fxint_layer_complete_vote(nominations, mf)
}
\arguments{
\item{data}{the data with the metaframe}

\item{mapping}{the aesthetic mapping}
}
\description{
This function generates the function to add an arbitrary number of layers
that depend on all specified layers.
}
\details{
\code{fxint_layer_complete} is driven by three extendible functions (see
\href{here}{?fxext_layer_complete_nominate()}):
\itemize{
\item \code{fxext_layer_complete_nominate} gathers the nominations
\item \code{fxext_layer_complete_veto} collects the vetoes
\item \code{fxext_layer_complete_vote} collects the votes
}
}
\section{Functions}{
\itemize{
\item \code{fxint_layer_complete_nominate}: Gathers the nominations.

\item \code{fxint_layer_complete_vote}: vetoes (via \code{fxext_layer_complete_veto})
and then votes (via \code{fxext_layer_complete_vote}) for the nominations
}}

