% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effex-ggplot-s4-fxName.R
\docType{class}
\name{fxName-class}
\alias{fxName-class}
\alias{.fxName}
\alias{indicatorsFxName-class}
\alias{.indicatorsFxName}
\alias{geomFxName-class}
\alias{.geomFxName}
\alias{descriptionFxName-class}
\alias{.descriptionFxName}
\alias{fxName}
\title{fxName class}
\usage{
fxName(fx)
}
\arguments{
\item{fx}{The fx If a subclass "<fx>FxName" exists, this will be
instantiated.}
}
\description{
In order to allow for flexible inheritance of additive functions, we define
a dummy class for every fx name (see also \code{\link[=aesName]{aesName()}}). The available methods
are:

\itemize{
\item indicators
}

\itemize{
\item geom
}

\itemize{
\item description
}
}
\examples{
fxName(fx = "geom")
fxName(fx = "description")
fxName(fx = "x")
}
\seealso{
aesName
}
