% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxPromise.R
\docType{class}
\name{fxp-class}
\alias{fxp-class}
\alias{createFx}
\title{Effex-Promises}
\usage{
createFx(fxp, indicator)
}
\description{
This class describes how to generate a certain fx object from
the indicator dataframe. The function \code{createFx} then yields the \code{fx} object.
}
