% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxDescription.R
\docType{class}
\name{fxpDescription-class}
\alias{fxpDescription-class}
\alias{.fxpDescription}
\alias{fxpDescription}
\alias{createFx,fxpDescription-method}
\alias{fxDescription}
\title{Effex Class: Descriptions}
\usage{
fxpDescription(...)

\S4method{createFx}{fxpDescription}(fxp, indicator)

fxDescription(tbl)
}
\arguments{
\item{...}{The different elements as mappings.}

\item{tbl}{The tibble.}
}
\value{
a data frame of the different elements
}
\description{
This effective and explicit class contains any form of descriptions. It is a
mapping of  A number of special fields is listed below:
\itemize{
\item label: a label for the indicator, e. g. "Number of cylinders"
}
}
\examples{
fxpDescription(labels = labels,
  description = purrr::map_chr(paste0("descs/", labels),
                               ~ readr::read_file(.))
)
}
