% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxGeom.R
\docType{class}
\name{fxpGeom-class}
\alias{fxpGeom-class}
\alias{.fxpGeom}
\alias{fxpGeom}
\alias{createFx,fxpGeom-method}
\title{Effex Promise: Geom}
\usage{
fxpGeom(mapping)

\S4method{createFx}{fxpGeom}(fxp, indicator)
}
\arguments{
\item{mapping}{A function that turns one row of the indicator data frame into
an \code{fxGeom} object. Must have as its only argument \code{indicator_row}.}
}
\description{
This class creates \code{fxGeom}.
}
