% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-fxGeomContinuous.R
\name{get_alpha}
\alias{get_alpha}
\title{Transparency for large scatter plots}
\usage{
get_alpha(fxGeom_alpha.threshold, fxGeom_alpha.half, fxGeom_alpha.min,
  n_row)
}
\arguments{
\item{fxGeom_alpha.threshold}{What is the threshold number of data points to
let transparency increase? Default is 1000}

\item{fxGeom_alpha.half}{What number of data points should result in
\code{alpha = 0.5}? Default is 3000}

\item{fxGeom_alpha.min}{What is the minimal \code{alpha}? Default is 0.05.}
}
\description{
This function computes \code{alpha} for scatter plots with a large number of data
points. It is intended to alleviate overplotting.
}
\keyword{internal}
