% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture-set.R
\name{get_architecture}
\alias{get_architecture}
\alias{set_architecture.default}
\alias{set_architecture.architecture}
\alias{set_architecture.NULL}
\title{Standard architecture}
\usage{
get_architecture()

\method{set_architecture}{default}(archie = NULL)

\method{set_architecture}{architecture}(archie = NULL)

\method{set_architecture}{NULL}(archie = NULL)
}
\description{
The \code{set_architecture}/\code{get_architecture} framework allows you to use
architectures more effectively without having to supply the architecture of
interest every time.
}
\section{Functions}{
\itemize{
\item \code{get_architecture}: What is the standard architecture?
}}

