% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effex-ggplot.R
\name{get_inds}
\alias{get_inds}
\title{Get the indicator names of an aesthetic}
\usage{
get_inds(mapping)
}
\description{
This function will later be extended such that it encompasses selection of
several indicators.
}
