% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-class.R
\name{get_ref_edges}
\alias{get_ref_edges}
\title{Let edges reference the relations and variables}
\usage{
get_ref_edges(variables, relations, edges)
}
\description{
Let edges reference the relations and variables
}
