% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable.R
\name{get_variable}
\alias{get_variable}
\title{Get Variable}
\usage{
get_variable(collection, indicator_names, internal = TRUE, ...)
}
\arguments{
\item{collection}{The collection of variables that have already been
imported. On the one hand, this is a shortcut as it automatically joins
the different collections. On the other hand, variables like a regression
result make use of existing variables. Collections can only be regarded in
\code{get_variable}.}

\item{indicator_names}{The names of the new indicators.}

\item{internal}{If existing, should the data be read in as internal to R? If
\code{internal} is TRUE but \code{get_variable_internal} fails, a warning is printed.}

\item{...}{additional arguments to \code{get_raw_variable} resp.
\code{get_variable_internal}.}
}
\description{
Get a variable from a list of indicators.
}
