% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{ind_source}
\alias{ind_source}
\alias{ind_*}
\title{Indicator Access Functions}
\usage{
ind_source(indicator)
}
\arguments{
\item{indicator}{The indicator or indicator element we are interested in.}
}
\value{
The sources of an indicator or an indicator element.
}
\description{
These functions provide access to information on the indicators.
}
