% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\docType{class}
\name{Indicator-class}
\alias{Indicator-class}
\title{Indicators}
\description{
Every element of a \code{collection_df} contains the class \code{indicator}. It
therefore has the minimal variable structure. Besides \code{name}, the slots
are optional and introduce some convenience.
}
\section{Slots}{

\describe{
\item{\code{ind_source}}{\code{IndSource}. The source which is used to read in variables,
see \linkS4class{IndSource}.}

\item{\code{name}}{\code{character}. The name of the corresponding column in the
\code{collection_df}}

\item{\code{raw_name}}{\code{character}. Gives the corresponding name in the raw variable.
Defaults to \code{name}.}

\item{\code{aliases}}{\code{character}. Alternative names.}

\item{\code{internal}}{\code{logical}. Should the prepackaged data be used via
\code{\link[=get_variable_internal]{get_variable_internal()}}?}
}}

