% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicator}
\alias{indicator}
\title{Indicators}
\usage{
indicator(ind_name = character(0), present = TRUE, ...)
}
\description{
Indicators describe the variables of a collection. They can be specified or
are added automatically.
}
