% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\docType{class}
\name{IndSource-class}
\alias{IndSource-class}
\alias{initialize,IndSource-method}
\title{Indicator Sources}
\usage{
\S4method{initialize}{IndSource}(.Object, file = character(0), ...)
}
\description{
Reading in and Generating variables uses the class \code{IndSource}. Use this
class to define a new method for \code{\link[=get_raw_variable]{get_raw_variable()}}. It may depend on
parameters like files or formulas.
}
\section{Slots}{

\describe{
\item{\code{file}}{\code{character}. The filename or URL for external data. Not necessary
/ set to \code{character(0)}, if the indicator only works with internal data.
\code{\link[=get_raw_variable]{get_raw_variable()}} should work with any allowed file argument. Neither
\code{\link[=tidy]{tidy()}} nor \code{\link[=clean]{clean()}} should depend on \code{file}.}
}}

