% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{is_architecture}
\alias{is_architecture}
\alias{is_verbose}
\alias{set_verbose}
\title{Architecture Infos}
\usage{
is_architecture(x)

is_verbose(archie)

set_verbose(archie, verbose)
}
\arguments{
\item{x}{a possible architecture}

\item{archie}{an architecture}

\item{verbose}{logical}
}
\description{
These functions provide additional info about an architecture.
}
\section{Functions}{
\itemize{
\item \code{is_architecture}: Is x an architecture?

\item \code{is_verbose}: Is the architecture verbose?

\item \code{set_verbose}: Change the architecture's verbosity.
}}

