% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-class.R
\name{is_variable}
\alias{is_variable}
\title{Variables}
\usage{
is_variable(x)
}
\arguments{
\item{x}{a possible variable}
}
\description{
Variables are can be defined by their content or operation. This constitutes
the two subclasses.
}
\details{
If you want to define a variable, you need to specify which subclass you have
by using the corresponding constructor function.
}
