% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-join.R
\name{join}
\alias{join}
\alias{join.collection_df}
\alias{join.collection}
\title{Join Collections}
\usage{
join(x, y, copy = FALSE, ...)

\method{join}{collection_df}(x, y, copy = FALSE, ...)

\method{join}{collection}(x, y, copy = FALSE, name = NULL, ...)
}
\arguments{
\item{x}{a \code{collection_df}.}

\item{y}{a \code{collection_df} with identical keys as \code{x}}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{...}{other parameters passed onto methods, for instance, \code{na_matches}
to control how \code{NA} values are matched.  See \link{join.tbl_df} for more.}

\item{name}{a name for \code{y}. If unspecified, \code{join} sets the name empty (i. e.
""). If \code{y} is a collection \code{name = NULL} does not affect the names but
\code{name = ""} does.}
}
\description{
\code{join} is a canonical form of the \code{*_join} set of functions. It has no \code{by}
argument as it joins by the keys of \code{x} and \code{y} and only allows identical
values of identical variables for identical keys.
}
\section{Methods (by class)}{
\itemize{
\item \code{collection}: Joins a \code{collection} with a \code{collection} or a
\code{collection_df}. If \code{y} is a \code{collection_df} a \code{name} for \code{y} can be
specified.
}}

