% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-write-json.R
\name{json_file_data}
\alias{json_file_data}
\title{Generate the columns .folder_path and .file_path}
\usage{
json_file_data(data, path)
}
\description{
This function generates the paths for a data frame. It returns a data frame
that is not grouped by any variables except \code{.file_path}. It also cleans the
potential file names (see \code{\link[=save_filenames]{save_filenames()}})
}
