% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-elements.R
\name{key}
\alias{key}
\alias{key.collection_df}
\alias{key.function}
\alias{key.default}
\alias{change_key.collection_df}
\alias{change_key.function}
\alias{change_key.default}
\title{Keys}
\usage{
key(x)

\method{key}{collection_df}(x)

\method{key}{function}(x)

\method{key}{default}(x)

\method{change_key}{collection_df}(x, key)

\method{change_key}{function}(x, key)

\method{change_key}{default}(x, key)
}
\arguments{
\item{x}{an R object}
}
\value{
The keys of x.
}
\description{
Keys are defined as columns of a table resp. arguments of a function that
uniquely determine specific observations resp. the function's return. They
simplify management of a collection as the canonical way of splitting up data
frames.
}
\section{Methods (by class)}{
\itemize{
\item \code{collection_df}: 

\item \code{function}: 

\item \code{default}: 

\item \code{collection_df}: 

\item \code{function}: 

\item \code{default}: 
}}

