% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key_to_char}
\alias{key_to_char}
\title{Key as character}
\usage{
key_to_char(x)
}
\arguments{
\item{x}{the key, as \code{key} or as \code{character}.}
}
\description{
For the purpose of joining columns, keys may be transformed to \code{character}
vectors. For this reason, ", " may currently not be used in a column name.
I do not expect this to be a problem. Should one arise, I will change the
separating characters.
}
