% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp-set.R
\name{match_els}
\alias{match_els}
\alias{\%el\%}
\title{Set operations}
\usage{
match_els(x, S, nomatch = 0)

x \%el\% S
}
\arguments{
\item{x}{a possible element}

\item{S}{a set.}
}
\description{
Supplementary set operations
}
\examples{
match_els(Content("Intelligence"),
list(Content("Intelligence"), 2, Content("uzwcjgh"))
}
