% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaframe.R
\name{as_metaframe}
\alias{as_metaframe}
\alias{is_metaframe}
\alias{metaframe}
\alias{set_metaframe}
\alias{metaframe<-}
\alias{has_metaframe}
\alias{mutate_mf}
\title{Metaframe of data}
\usage{
as_metaframe(df)

is_metaframe(x)

metaframe(x)

set_metaframe(x, value)

metaframe(x) <- value

has_metaframe(x)

mutate_mf(data, ...)
}
\arguments{
\item{df}{a data frame that contains information on indicators}

\item{x}{an object}

\item{value}{a metaframe}

\item{x}{an object}
}
\description{
Get and set the \code{\link[=metaframe]{metaframe()}} of certain data.
}
\section{Functions}{
\itemize{
\item \code{has_metaframe}: determines whether x has a metaframe

\item \code{mutate_mf}: allows you to modify the metaframe by accessing the
data itself. If the metaframe does not exist, it creates one.
}}

