% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{namekey}
\alias{namekey}
\alias{names.collection}
\alias{key.collection}
\alias{names<-.collection}
\alias{change_key.collection}
\alias{change_namekey}
\alias{namekey<-}
\title{Attributes of Collections}
\usage{
namekey(x)

\method{names}{collection}(x)

\method{key}{collection}(x)

\method{names}{collection}(x) <- value

\method{change_key}{collection}(x, key)

change_namekey(x, value)

namekey(x) <- value
}
\description{
Names of collections are only unique in combination with their keys. In many
cases, it is sensible to avoid names and only use keys.
}
