% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\docType{class}
\name{NestedIndSource-class}
\alias{NestedIndSource-class}
\alias{initialize,NestedIndSource-method}
\title{Nested Indicator Source}
\usage{
\S4method{initialize}{NestedIndSource}(.Object, ind_source, names, ...)
}
\description{
This object is created by \code{\link[=nest_IndSource]{nest_IndSource()}}. It specifies the names that
correspond to a specific indicator source.
}
