% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath.R
\name{new_filepath}
\alias{new_filepath}
\alias{fields}
\alias{reader}
\alias{is_filepath}
\alias{read_filepath}
\alias{as_filepath}
\alias{[.filepath}
\title{Filepaths}
\usage{
new_filepath(path, fields, reader)

fields(x)

reader(x)

is_filepath(x)

read_filepath(filepath, ...)

as_filepath(x)

\method{[}{filepath}(x, i)
}
\arguments{
\item{path}{The path to the file}

\item{fields}{The fields within the file}

\item{reader}{The reader of the files. A function that takes \code{file} for the
filename and optionally additional arguments. For compatibility, it is
recommended to allow dots arguments.}
}
\description{
Filepaths keep metaframes manageable by externalizing semantic information to
files and capturing their columns with the attribute field.
}
\section{Functions}{
\itemize{
\item \code{read_filepath}: takes a filepath and uses the reader-function to
read all files in as a data frame. It can be identified with the supplied
filepath by the \code{name}-column.
}}

