% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-obj-infer.R
\name{new_FxInfer}
\alias{new_FxInfer}
\alias{is_FxInfer}
\alias{subclass}
\title{FxInfer: S3 class for inferring columns via \code{\link[=fx_infer]{fx_infer()}}}
\usage{
new_FxInfer(subclass = NULL)

is_FxInfer(x)

subclass(FxInfer)
}
\arguments{
\item{subclass}{The class of the object.}

\item{x}{an object.}

\item{FxInfer}{an \code{FxInfer}-object.}
}
\value{
An empty list of class \code{c(<subclass>FxInfer, "FxInfer")}
}
\description{
\code{new_FxInfer} creates a very simple class that only works via its class
attribute.
}
\section{Functions}{
\itemize{
\item \code{subclass}: returns the specified subclass
}}

