% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{new_key}
\alias{new_key}
\alias{as_key}
\alias{format.key}
\alias{print.key}
\title{Keys}
\usage{
new_key(...)

as_key(x)

\method{format}{key}(x, ...)

\method{print}{key}(x)
}
\arguments{
\item{...}{several character vectors.}
}
\description{
Keys are an essential component of \code{\link[=collection_df]{collection_df()}}s. They are a list of
character vectors and can therefore use a few extra operations that are
undefined.
}
