% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaframe.R
\name{new_metaframe}
\alias{new_metaframe}
\title{Metaframe}
\usage{
new_metaframe(...)
}
\description{
The metaframe is the place in which the effex information is maintained. It
consists of columns with exclusively reserved names many of which are used by
certain effex functions. These names are documented in the respective
functions. Required column names for any metaframe are:
}
\details{
\itemize{
\item \code{name}: identify the column name
}
}
