% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-nomination.R
\name{nom_access}
\alias{nom_access}
\alias{nom_layers}
\alias{nom_facets}
\alias{nom_scales}
\alias{nom_coords}
\alias{nom_others}
\title{Access functions to a nomination}
\usage{
nom_layers(nomination)

nom_facets(nomination)

nom_scales(nomination)

nom_coords(nomination)

nom_others(nomination)
}
\description{
These functions provide access to resp. change the different kinds of layers
of a nomination.
}
\section{Functions}{
\itemize{
\item \code{nom_layers}: all layers (class "Layer")

\item \code{nom_facets}: all facets (class "Facet")

\item \code{nom_scales}: all scales (class "Scale")

\item \code{nom_coords}: all coordinate systems (class "Coord")

\item \code{nom_others}: all other ggproto objects
}}

