% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-nomination.R
\name{nomination}
\alias{nomination}
\alias{add_to_nom.Layer}
\alias{add_to_nom.Facet}
\alias{add_to_nom.Scale}
\alias{add_to_nom.Coord}
\alias{add_to_nom.ggproto}
\alias{is_nomination}
\title{Nominations}
\usage{
nomination(...)

\method{add_to_nom}{Layer}(nomination_el, nomination)

\method{add_to_nom}{Layer}(nomination_el, nomination)

\method{add_to_nom}{Facet}(nomination_el, nomination)

\method{add_to_nom}{Scale}(nomination_el, nomination)

\method{add_to_nom}{Coord}(nomination_el, nomination)

\method{add_to_nom}{ggproto}(nomination_el, nomination)

is_nomination(x)
}
\arguments{
\item{...}{Layers that are nominated together or nominations that are
concatenated}

\item{x}{an object}
}
\description{
\itemize{
\item layers
\item scales
\item faceting specifications
\item coordinate systems
\item other ggproto objects
}

This function is intended to be extendible via the generic \code{add_to_nom}.
}
\details{
Nominations are essential objects in the construction of
\code{\link[=fx_ggplot]{fx_ggplot()}}. Their structure is a list of lists where each list contains a
different kind of "ggproto" object, namely:
}
