% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-helpers.R
\name{null_helper}
\alias{null_helper}
\title{NULL helper}
\usage{
null_helper(value, default)
}
\description{
This function is a simple helper that replaces \code{list(NULL)} with \code{NULL} and
\code{NULL} with \code{default}. It allows an easier passing of arguments in the
external functions.
}
