% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-methods.R
\name{print.content}
\alias{print.content}
\title{Print Variables}
\usage{
\method{print}{content}(x, parse = 5, ...)
}
\arguments{
\item{parse}{number of contents that are printed or logical \code{FALSE} that
indicates that all contents are to be printed.}
}
\description{
\code{print} provides a clearer approach to a \code{variable}.
\code{print.content} prints the title of the character and a parsed definition as
well as the first five attributes.
}
\examples{
Content(title = "test", definition = "We are conducting a test",
        supp = "there is another attribute that is too long to be
                displayed in one line. We therefore will parse it to the
                appropriate length.")

}
