% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp-print.R
\name{print_in_lines}
\alias{print_in_lines}
\alias{print_in_lines_input}
\title{Supplementary print functions}
\usage{
print_in_lines(x, nchars_per_line = NULL, nlines = 1, nchars = NULL,
  linebreak = "\\n", parsesign = "...")

print_in_lines_input(x, nchars_per_line = NULL, nlines = 1, nchars = NULL)
}
\arguments{
\item{x}{character or coerced to character.}

\item{nchars_per_line}{number of characters in one line}

\item{nlines}{number of lines to be printed}

\item{nchars}{number of characters to be printed}

\item{linebreak}{what should be used as the breaking character? Default is
\code{"\n"}.}

\item{parsesign}{how should the function signal a parsed character? Default
is \code{"..."}. Should not be longer than a few characters and
can be omitted with \code{""}.}
}
\description{
Help with printing different classes. \code{print_in_lines()} generates a
character, in a specified number of lines with specified characters per line.
Print with \code{cat()}.

\code{print_in_lines_input()} assures appropriate input of the
numeric specifications.
}
\examples{
cat(print_in_lines("Weird lyrics with weird linebreak",
                   nchars_per_line = 5,
                   nlines = 4))

}
