% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-helpers.R
\name{required_aes}
\alias{required_aes}
\title{Required aesthetics of a layer}
\usage{
required_aes(layer)
}
\arguments{
\item{layer}{a layer.}
}
\description{
Concatenates the required aesthetics of the geom, stats and position of a
layer.
}
\details{
This function is not very helpful because the required aesthetics are
overestimated, see \code{GeomBoxplot}. It will thus not be employed for now.
}
\seealso{
\code{\link[ggplot2:layer]{ggplot2::layer()}}
}
