% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-write-json.R
\name{save_filenames}
\alias{save_filenames}
\title{Change filenames so that they are suitable for file naming}
\usage{
save_filenames(filenames)
}
\description{
Removes or replaces characters that cannot name a file or folder.
}
