% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope-class.R
\name{Scope}
\alias{Scope}
\title{\code{Scope}}
\usage{
Scope(name, slots, supscope = "Compound")
}
\arguments{
\item{name}{a \code{character}. Specifies the name of the scope.}

\item{slots}{a named vector. Specifies the additional slots the scope needs.}

\item{supscope}{\code{Scopes} may be stacked by providing a supscope.}
}
\description{
\code{Scope}s define the level on which we make observation about our \code{Compound}s.
They are referenced in \code{Compound}s as well as many \code{Operator}s.
}
\seealso{
\link{Compound}
}
