% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp-set.R
\name{Set}
\alias{Set}
\alias{Set_list}
\title{Sets}
\usage{
Set(...)

Set_list(list_inp)
}
\arguments{
\item{...}{the elements of the set.}

\item{list_inp}{the elements of the set as a list.}
}
\description{
Creates a list with class \code{Set} from the input which contains unique
elements.
}
