% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-info-statistics.R
\name{stat_names}
\alias{stat_names}
\title{Names of statistics}
\usage{
stat_names(statistics)
}
\description{
Get the names of the \code{statistics} in \code{\link[=fxe_info.fxd_info_stats]{fxe_info.fxd_info_stats()}}.
}
