% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-class.R
\name{Structure}
\alias{Structure}
\alias{Structure.default}
\alias{Structure.variable}
\title{Structure}
\usage{
Structure(variables, ...)

\method{Structure}{default}(variables, ...)

\method{Structure}{variable}(variables, relations = NULL, edges)
}
\arguments{
\item{variables}{The variables of the structure.}
}
\description{
Structures are collections of variables and relations between them -
essentially graphs.
}
