% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tectr.R
\docType{package}
\name{tectr-package}
\alias{tectr}
\alias{tectr-package}
\title{tectr}
\description{
tectr
}
\author{
\strong{Maintainer}: Samuel Lippl \email{sfc.lippl@gmail.com}

Authors:
\itemize{
  \item Fabian Obster \email{fabian.obster@gmail.com}
}

}
