% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{type_sum.key}
\alias{type_sum.key}
\title{#' @importFrom pillar is_vector_s3
#' @export}
\usage{
\method{type_sum}{key}(x)
}
\description{
is_vector_s3.key <- function(x) TRUE
}
\details{
#' @importFrom pillar obj_sum
#' @export

obj_sum.key <- function(x) rep("key", length(x))
}
