% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-source.R
\name{variable_source}
\alias{variable_source}
\title{Source Variables}
\usage{
variable_source(source_fun, file = NULL, ...)
}
\arguments{
\item{source_fun}{the function that reads in the source. May or may not be
dependent on the file parameter.}

\item{file}{optional file that is the source. See details}

\item{...}{variable attributes}
}
\description{
Source Variables do not depend on other variables. They basically initialize
the data frame by reading in the external source or generating data in
another way that is independent from other variables.
}
\details{
When applicable, it is recommended to use the file parameter, as it
makes the variable's structure clearer
}
