% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-visualization.R
\name{visualize_struct}
\alias{visualize_struct}
\title{Visualize a structure graph}
\usage{
visualize_struct(struct)
}
\description{
For now, this function is really inflexible because it is only intended for
demonstration purposes in the shiny app.
}
