% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collection}
\alias{collection}
\title{Collections}
\usage{
collection(...)
}
\arguments{
\item{...}{objects of the class \code{collection_df}}
}
\description{
Collections are composed of collection elements (at the time only data
frames). Each collection can be uniquely identified by a key and/or a name,
where every unique combination of key and name provides a \code{collection_el}
with the respective key.
}
