% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-class.R
\name{Content}
\alias{Content}
\alias{.Content}
\alias{is_content}
\title{Content}
\usage{
Content(title, definition = "", ...)

.Content(title, definition = "", ...)

is_content(x)
}
\arguments{
\item{title}{the variable name}

\item{definition}{a short definition of the variable}

\item{...}{details on the content}
}
\description{
\code{content} is a \code{variable} that is defined non-statistically. It therefore
consists of a character vector carrying the definition and possible comments
carried as attributes.
}
\details{
Do not set \code{content} as \code{class(x) <- "content"} but only with the constructor
function.
}
\examples{
int <- Content(title = "Intelligence",
               definition = "Intelligence has been defined in many different
                             ways including as one's capacity for logic,
                             understanding, self-awareness, learning,
                             emotional knowledge, planning, creativity, and
                             problem solving. It can be more generally
                             described as the ability or inclination to
                             perceive or deduce information, and to retain it
                              as knowledge to be applied towards adaptive
                             behaviors within an environment or context.",
                             Source = "Wikipedia",
                             Subject = "Psychology")

}
