% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-methods.R
\name{extend}
\alias{extend}
\alias{extend_variables}
\alias{extend_relations}
\title{Extend a structure}
\usage{
extend(structure, variables = NULL, relations = NULL, edges = NULL)

extend_variables(structure, variables)

extend_relations(structure, relations)
}
\arguments{
\item{structure}{a structure to be extended}

\item{variables}{The variables of the structure.}

\item{relations}{relations that extend the structure}

\item{edges}{new connection between contents and relations}
}
\description{
Extend a structure with new content, new relations or new edges.
}
