% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-default.R
\name{fx_default}
\alias{fx_default}
\alias{fxi_default}
\alias{fxe_default}
\alias{fxe_default.fxd_default}
\title{Effex functions: Defaults for the metaframe}
\usage{
fx_default(data, columns = character(0), ..., overwrite = FALSE,
  mustWork = TRUE)

fxi_default(data, col, ...)

fxe_default(data, mf, col, ..., mustWork = TRUE)

\method{fxe_default}{fxd_default}(data, mf, col, ..., mustWork = TRUE)
}
\arguments{
\item{data}{The dataframe with a metaframe attribute}

\item{columns}{\code{character}. The columns of the metaframe that should be
provided.}

\item{...}{Parameters to give on to the methods of \code{\link[=fxe_default]{fxe_default()}}}

\item{overwrite}{\code{logical}. Should existing columns be overwritten? If \code{TRUE}
might also delete columns without compatible method.}

\item{mustWork}{If no compatible method is found, should the method return
\code{NULL} or throw an error?}
}
\description{
The \code{fx_default}-family consists of functions with a consistent interface that
that sets a default value for the specified metaframe column. Their intention
is an explicit standard mode
which allows to concentrate on changing special cases only.
\code{fx_default} itself wraps around these functions by modifying the metaframe
for any number of metaframe columns. This is, of course only useful if
modifications need not be applied.
}
\details{
\code{fx_default} can be extended by adding a method to the generic \code{fxe_default}.
It is recommended to specify the particular method following the scheme
\code{fx_default_<column name>} to make it easier to use effex default outside an
entire data frame. To achieve that \code{fxi_default} creates a structure of class
<column name> with superclass "fx_default" via \code{\link[=fxd]{fxd()}}
}
