% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot.R
\name{fx_ggplot}
\alias{fx_ggplot}
\title{Effex Function: Visualization}
\usage{
fx_ggplot(data, mapping)
}
\arguments{
\item{data}{A dataframe with \code{\link[=metaframe]{metaframe()}} as an attribute}

\item{mapping}{An aesthetic mapping}
}
\description{
This function creates an effectively and explicitly specified visualization
of several variables in line with the syntax of \link{ggplot2}.
}
\details{
Required columns in the metaframe:
\itemize{
\item \code{fxGeom_class}: What class does the geometry have?
}

Columns that are used:
\itemize{
\item \code{fxGeom_nominations}: A tibble of nominated layers (see
\code{\link[=fxint_layer_complete]{fxint_layer_complete()}})
\item \code{fxGeom_vetos}: A function that returns the nominations data frame without
the vetoed layers
\item \code{fxGeom_votes}: A function that returns the nominations data frame with the
new votes
}
}
