% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-info-statistics.R,
%   R/fx-fun-info-title.R, R/fx-fun-info.R
\name{fxe_info.fxd_info_stats}
\alias{fxe_info.fxd_info_stats}
\alias{fxe_info.fxd_info_title}
\alias{fx_info}
\alias{fxe_info}
\alias{fxe_info.fxd_info}
\title{Effex Function: Information}
\usage{
\method{fxe_info}{fxd_info_stats}(data, topic, statistics, ...)

\method{fxe_info}{fxd_info_title}(data, topic, ...)

fx_info(data, topic, ...)

fxe_info(data, topic, ...)

\method{fxe_info}{fxd_info}(data, topic, ...)
}
\arguments{
\item{data}{A dataframe, possibly with a metaframe}

\item{topic}{The topic of the information}

\item{statistics}{a character vector or a list of characters and functions.
Function slots must have a name. Functions can accept arguments from
metaframe columns. The corresponding column will be fed in first and there
will be an additional data column with the entire data. All arguments but
the column will only be fed into functions, not characters.}

\item{fxInfo_title}{\code{character}, \code{NULL} or list of both. If}

\item{arguments}{for the information methods}
}
\description{
This function provides versatile information on the different variables.
\code{\link[=fx_output]{fx_output()}} turns this information into a string.
}
\section{Methods (by generic)}{
\itemize{
\item \code{fxe_info}: provides a statistical summary table

\item \code{fxe_info}: determines a title via \code{\link[=fx_info_title]{fx_info_title()}}

\item \code{fxe_info}: If no method for this topic is supplied, this function
looks up whether a column \code{fxInfo_<column>} exists.
}}

\section{Extension Mechanism}{

An extension may be defined via \code{fxe_info()} which dispatches over
\code{fxd("info", topic)}.
}

