% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R
\docType{methods}
\name{fxe_labeller}
\alias{fxe_labeller}
\alias{fxe_labeller,fxGeomDiscrete-method}
\title{Effex Extendibles in \code{fx_ggplot}: facetting labellers}
\usage{
fxe_labeller(fxGeom_class, name, ...)

\S4method{fxe_labeller}{fxGeomDiscrete}(fxGeom_class, name, fxGeom_limits,
  fxGeom_breaks = NULL, ...)
}
\value{
a valid argument to \code{\link[ggplot2:labeller]{ggplot2::labeller()}}. Supplies \code{\link[=fx_ggplot]{fx_ggplot()}}
with the labels for the facetting variable
}
\description{
Effex Extendibles in \code{fx_ggplot}: facetting labellers
}
\section{Methods (by class)}{
\itemize{
\item \code{fxGeomDiscrete}: If the breaks are specified, these are used for the
labeller. If not the default argument \code{\link[ggplot2:label_value]{ggplot2::label_value()}} is returned.
}}

