% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R,
%   R/fx-fun-ggplot-fxGeomContinuous.R, R/fx-fun-ggplot-fxGeomDiscrete.R,
%   R/fx-fun-ggplot-fxGeomOrdinalCI.R, R/fx-fun-ggplot-fxGeomSpatial.R,
%   R/fx-fun-ggplot-fxGeomTime.R
\docType{methods}
\name{fxe_layer_complete_nominate}
\alias{fxe_layer_complete_nominate}
\alias{fxe_layer_complete_nominate,fxGeom,AesName-method}
\alias{fxe_layer_complete_veto}
\alias{fxe_layer_complete_nominate,fxGeomContinuous,xAesName-method}
\alias{fxe_layer_complete_nominate,fxGeomContinuous,yAesName-method}
\alias{fxe_layer_complete_nominate,fxGeomDiscrete,xAesName-method}
\alias{fxe_layer_complete_nominate,fxGeomDiscrete,yAesName-method}
\alias{fxe_layer_complete_nominate,fxGeomOrdinalCI,yAesName-method}
\alias{fxe_layer_complete_nominate,fxGeomSpatial,geomAesName-method}
\alias{fxe_layer_complete_nominate,fxGeomTime,xAesName-method}
\title{Effex Extendibles in \code{fx_ggplot}: nominate layers}
\usage{
fxe_layer_complete_nominate(fx_geom, aes_name, data, ...)

\S4method{fxe_layer_complete_nominate}{fxGeom,AesName}(fx_geom, aes_name,
  data, ..., fxGeom_nominations = NULL)

fxe_layer_complete_veto(nomination, fx_geom, aes_name, data, ...)

\S4method{fxe_layer_complete_nominate}{fxGeomContinuous,xAesName}(fx_geom,
  aes_name, data, ..., name, fxGeom_nominations = NULL,
  fxGeom_alpha.threshold = NULL, fxGeom_alpha.half = NULL,
  fxGeom_alpha.min = NULL)

\S4method{fxe_layer_complete_nominate}{fxGeomContinuous,yAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL,
  fxGeom_alpha.threshold = NULL, fxGeom_alpha.half = NULL,
  fxGeom_alpha.min = NULL)

\S4method{fxe_layer_complete_nominate}{fxGeomDiscrete,xAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL)

\S4method{fxe_layer_complete_nominate}{fxGeomDiscrete,yAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL)

\S4method{fxe_layer_complete_nominate}{fxGeomOrdinalCI,yAesName}(fx_geom,
  aes_name, data, ..., name, fxGeom_nominations = NULL,
  fxGeom_assoc_vars = NULL, fxGeom_errorbar.threshold = NULL)

\S4method{fxe_layer_complete_nominate}{fxGeomSpatial,geomAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL)

\S4method{fxe_layer_complete_nominate}{fxGeomTime,xAesName}(fx_geom,
  aes_name, data, ..., name, fxGeom_nominations = NULL,
  fxGeom_alpha.threshold = NULL, fxGeom_alpha.half = NULL,
  fxGeom_alpha.min = NULL, fxGeom_assoc_vars = NULL,
  fxGeom_shadow.threshold = NULL, fxGeom_shadow.alpha = NULL)
}
\arguments{
\item{fx_geom}{An S4 class argument which inherits from \linkS4class{fxGeom}.}

\item{aes_name}{An S4 class argument which inherits from \linkS4class{AesName}.}

\item{fxGeom_nominations}{Allows nominating additional elements.}

\item{nomination}{the \code{\link[=nomination]{nomination()}} which is to be considered for a veto}

\item{fxGeom_alpha.threshold}{What is the threshold number of data points to
let transparency increase? Default is 1000}

\item{fxGeom_alpha.half}{What number of data points should result in
\code{alpha = 0.5}? Default is 3000}

\item{fxGeom_alpha.min}{What is the minimal \code{alpha}? Default is 0.05.}

\item{fxGeom_shadow.threshold}{What is the threshold for the number of groups
so that the different lines are greyed out and a smoother is fitted? Default
is 10}
}
\value{
\code{fxe_layer_complete_veto} returns a logical value that indicates whether
this layer should be vetoed, i. e. removed from the nominations list.
}
\description{
Effex Extendibles in \code{fx_ggplot}: nominate layers
}
\section{Methods (by class)}{
\itemize{
\item \code{fx_geom = fxGeom,aes_name = AesName}: returns the list of all nomination
which have been specified manually within the metaframe. If the argument is
\code{NULL}, the function returns an empty list.

\item \code{fx_geom = fxGeomContinuous,aes_name = xAesName}: \itemize{
\item a scatter plot whose transparency depends on the number of data points
(see \code{\link[=get_alpha]{get_alpha()}})
\itemize{
\item a histogram
\item a density plot
\item a hexagonal heatmap.
}
}

\item \code{fx_geom = fxGeomContinuous,aes_name = yAesName}: \itemize{
\item a boxplot
\itemize{
\item a scatter plot whose transparency depends on the number of data points
\item a hexagonal heatmap
}
}

\item \code{fx_geom = fxGeomDiscrete,aes_name = xAesName}: \itemize{
\item bar plot
\itemize{
\item heatmap
\item boxplot
}
}

\item \code{fx_geom = fxGeomDiscrete,aes_name = yAesName}: \itemize{
\item heatmap
}

\item \code{fx_geom = fxGeomOrdinalCI,aes_name = yAesName}: \itemize{
\item a line plot together with a shaded area for the confidence
interval
\itemize{
\item a scatter plot with uncertainty lines
}
}

\item \code{fx_geom = fxGeomSpatial,aes_name = geomAesName}: \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}}

\item \code{fx_geom = fxGeomTime,aes_name = xAesName}: \itemize{
\item a line plot that is possibly grouped according to the group aesthetic
in \code{fxGeom_assoc_vars}. If there are more than 10 groups, the individual
lines become transparent and a smooth function is fitted
\itemize{
\item a step plot with the same specifications
}
}
}}

\seealso{
\code{\link[=fxi_layer_complete]{fxi_layer_complete()}}
}
