% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R
\docType{methods}
\name{fxe_layer_single}
\alias{fxe_layer_single}
\alias{fxe_layer_single,fxGeom,AesName-method}
\title{Effex Extendibles: \code{fx-ggplot} - geometric information along one dimension}
\usage{
fxe_layer_single(fx_geom, aes_name, data, ...)

\S4method{fxe_layer_single}{fxGeom,AesName}(fx_geom, aes_name, data, ...)
}
\arguments{
\item{fx_geom}{the \code{\link[=fxGeom]{fxGeom()}} object}

\item{aes_name}{the \code{\link[=aesName]{aesName()}} object}
}
\description{
This is the extendible function for the single layer. It is dispatched
along the arguments \code{fx_geom} and \code{aes_name}. Arbitrary other arguments may
be added that should then be expected in the metaframe. It should always be
considered a possibility that these values are missing.
}
