% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R,
%   R/fx-fun-ggplot-fxGeomSpatial.R
\docType{methods}
\name{fxe_layer_single}
\alias{fxe_layer_single}
\alias{fxe_layer_single,fxGeom,AesName-method}
\alias{fxe_layer_single,fxGeomSpatial,AesName-method}
\title{Effex Extendibles in \code{fx-ggplot}: geometric information along one dimension}
\usage{
fxe_layer_single(fx_geom, aes_name, data, ...)

\S4method{fxe_layer_single}{fxGeom,AesName}(fx_geom, aes_name, data, ...)

\S4method{fxe_layer_single}{fxGeomSpatial,AesName}(fx_geom, aes_name, data,
  ...)
}
\arguments{
\item{fx_geom}{An S4 class argument which inherits from \linkS4class{fxGeom}.}

\item{aes_name}{An S4 class argument which inherits from \linkS4class{AesName}.}

\item{fx_geom}{the \code{\link[=fxGeom]{fxGeom()}} object}

\item{aes_name}{the \code{\link[=aesName]{aesName()}} object}
}
\description{
This is the extendible function for the independent elements. It is
dispatched along the S4 class arguments \code{fx_geom} and \code{aes_name}. Arbitrary
other parameters may
be added that should then be expected in the metaframe. It should always be
considered a possibility that these values are missing.
}
\section{Methods (by class)}{
\itemize{
\item \code{fx_geom = fxGeom,aes_name = AesName}: The standard mechanism simply adds the result of
\code{\link[=fxe_layer_scale]{fxe_layer_scale()}} and \code{\link[=fxe_layer_single_other]{fxe_layer_single_other()}} together.

\item \code{fx_geom = fxGeomSpatial,aes_name = AesName}: produces an error as the aesthetic must be
\code{geom}
}}

