% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R,
%   R/fx-fun-ggplot-fxGeomContinuous.R, R/fx-fun-ggplot-fxGeomDiscrete.R
\docType{methods}
\name{fxext_layer_complete_nominate}
\alias{fxext_layer_complete_nominate}
\alias{fxext_layer_complete_nominate,fxGeom,AesName-method}
\alias{fxext_layer_complete_veto}
\alias{fxext_layer_complete_vote}
\alias{fxext_layer_complete_nominate,fxGeomContinuous,xAesName-method}
\alias{fxext_layer_complete_nominate,fxGeomContinuous,yAesName-method}
\alias{fxext_layer_complete_nominate,fxGeomDiscrete,xAesName-method}
\alias{fxext_layer_complete_nominate,fxGeomDiscrete,yAesName-method}
\title{Effex Extendibles: \code{fx_ggplot} - nominate layers}
\usage{
fxext_layer_complete_nominate(fx_geom, aes_name, data, ...)

\S4method{fxext_layer_complete_nominate}{fxGeom,AesName}(fx_geom, aes_name,
  data, ..., fxGeom_nominations = NULL)

fxext_layer_complete_veto(nomination, fx_geom, aes_name, data, ...)

fxext_layer_complete_vote(nomination, fx_geom, aes_name, data, ...)


  \S4method{fxext_layer_complete_nominate}{fxGeomContinuous,xAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL)


  \S4method{fxext_layer_complete_nominate}{fxGeomContinuous,yAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL)

\S4method{fxext_layer_complete_nominate}{fxGeomDiscrete,xAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL)

\S4method{fxext_layer_complete_nominate}{fxGeomDiscrete,yAesName}(fx_geom,
  aes_name, data, ..., fxGeom_nominations = NULL)
}
\arguments{
\item{nomination}{a \code{\link[=nomination]{nomination()}}}
}
\description{
These functions drive \code{\link[=fxint_layer_complete]{fxint_layer_complete()}}. Provided that they fulfill
certain conditions they can be extended.
}
\details{
\code{fxext_layer_complete_nominate} returns a list of
\code{\link[=nomination]{nomination()}}(layer nominations). The default returns either an empty
list or, if provided the element of the metaframe column
\code{fxGeom_nominations}.

\code{fxext_layer_complete_veto} returns a logical value that indicates whether
this layer should be vetoed, i. e. removed from the nominations list. The
default method returns either \code{FALSE} or, if supplied with an element of the
metaframe column, it applies that function.
}
\seealso{
\code{\link[=fxint_layer_complete]{fxint_layer_complete()}}
}
