% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-fx-obj-fxGeom.R
\docType{class}
\name{fxGeom-class}
\alias{fxGeom-class}
\alias{.fxGeom}
\alias{fxGeomContinuous-class}
\alias{.fxGeomContinuous}
\alias{fxGeomContinuousCI-class}
\alias{.fxGeomContinuousCI}
\alias{fxGeomTime-class}
\alias{.fxGeomTime}
\alias{fxGeomDiscrete-class}
\alias{.fxGeomDiscrete}
\alias{fxGeomOrdinal-class}
\alias{.fxGeomOrdinal}
\alias{fxGeomOrdinalCI-class}
\alias{.fxGeomOrdinalCI}
\alias{fxGeomSpatial-class}
\alias{.fxGeomSpatial}
\alias{fxGeom}
\title{Effex Objects: fxGeom}
\usage{
fxGeom(geom_class = "")
}
\arguments{
\item{geom_class}{the subclass.}
}
\description{
fxGeom is a class that helps with method dispatch. Every subclass of type
<subclass> is instantiated by \code{fxGeom<subclass>} and can be created by
calling \code{fxGeom}. If the corresponding subclass does not exist, the function
will throw an error. The available subclasses are:

\itemize{
\item fxGeomContinuous
}

\preformatted{+ fxGeomContinuousCI
}

\preformatted{+ fxGeomTime
}

\itemize{
\item fxGeomCI
}

\itemize{
\item fxGeomDiscrete
}

\preformatted{+ fxGeomOrdinal
}

\preformatted{  + fxGeomOrdinalCI (also inherits from fxGeomOrdinal)
}

\itemize{
\item fxGeomSpatial
}
}
\examples{
fxGeom("Continuous")
fxGeom("Spatial")
fxGeom("")
}
