% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxGeom.R
\docType{class}
\name{fxGeom-class}
\alias{fxGeom-class}
\alias{.fxGeom}
\alias{fxContinuousGeom-class}
\alias{.fxContinuousGeom}
\alias{fxDiscreteGeom-class}
\alias{.fxDiscreteGeom}
\alias{fxContinuousGeom}
\alias{fxDiscreteGeom}
\title{Effex Class: Geom}
\usage{
fxContinuousGeom(limits = c(NA_real_, NA_real_))

fxDiscreteGeom(limits = NULL)
}
\description{
The class \code{fxGeom} contains different information on the properties, the
visualization of a certain variable should have. Its main subclasses are
\code{fxContinuousGeom} and \code{fxDiscreteGeom}.
}
\section{Slots}{

\describe{
\item{\code{limits}}{Range of a continuous variable. Default is \code{c(NA, NA)} as this
does not influence the limits.}

\item{\code{limits}}{\code{character}, \code{numeric} or \code{NULL}: Values of a discrete variable.
Default is NULL.}
}}

