% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot-aaa.R
\name{fxi_layer_complete}
\alias{fxi_layer_complete}
\title{Effex Internals in \code{fx_ggplot}: combine all different dimensions}
\usage{
fxi_layer_complete(data, mapping)
}
\arguments{
\item{data}{the data with the metaframe}

\item{mapping}{the aesthetic mapping}
}
\description{
This function generates the function to add an arbitrary number of elements
which depend on all specified layers.
}
\section{Voting System}{

This function employs a \emph{voting system} to determine a suitable
\code{\link[=nomination]{nomination()}} of elements.
\itemize{
\item \code{\link[=fxe_layer_nominate]{fxe_layer_nominate()}} gathers nominations
\item \code{\link[=fxe_layer_veto]{fxe_layer_veto()}} determines the vetos
\item \code{\link[=fxe_layer_vote]{fxe_layer_vote()}} gathers the votes
}

You can extend all three of these functions by S4 dispatch over the aesthetic
and the geometry class.
}

