% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx-fun-ggplot.R
\name{fxint_layer_complete}
\alias{fxint_layer_complete}
\title{Effex Internals: \code{fx_ggplot} - combine all different dimensions}
\usage{
fxint_layer_complete(data, mapping)
}
\arguments{
\item{data}{the data with the metaframe}

\item{mapping}{the aesthetic mapping}
}
\description{
\code{fxint_layer_complete} is mainly driven by two extendible functions:
\code{\link[=fxext_layer_complete_nominate]{fxext_layer_complete_nominate()}} which allows you to supply a data frame
with new nominations in the corresponding column for geom, stat, position and
params, and \code{\link[=fxext_layer_complete_vote]{fxext_layer_complete_vote()}} which allows you to give votes
depending on the geom, stat, position and params by modifying the \code{votes}
columns so that it adds your votes. If a certain combination is incompatible
with the data type, you may remove it.
}
