% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable.R
\name{get_variable}
\alias{get_variable}
\title{Get Variable}
\usage{
get_variable(indicator)
}
\arguments{
\item{indicator}{The indicators that are to be created.}
}
\value{
The final variable
}
\description{
\code{get_variable()} transforms indicators - promises of data - into variables,
i. e. data. You may specify the way create works by the functions
\code{get_raw_variable()}, \code{tidy()} and \code{clean()}.
}
