% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable.R
\name{get_variable}
\alias{get_variable}
\alias{get_raw_variable}
\alias{tidy}
\alias{clean}
\title{Get variable}
\usage{
get_variable(collection, indicators)

get_raw_variable(indicators, collection)

tidy(raw_variable, indicators)

clean(x, indicator = NULL, ...)
}
\description{
\code{get_variable} provides a framework for reading in indicators.
}
