% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicator}
\alias{indicator}
\alias{as_indicator}
\title{Indicator}
\usage{
indicator(name = NULL, source = "", ..., df = NULL)

as_indicator(spec = NULL, name = NULL, source = "", ...)
}
\arguments{
\item{name}{What is the indicator's name? Within each source, names have to
be unique.}

\item{source}{What is the indicator's source? If you create the indicator
from other indicators, the source is "composite". You may also specify it by
"" or by not mentioning source at all.}

\item{...}{Further attributes.}

\item{df}{You may also provide the information by a data frame. Name and
source must be a part of \code{df} in that case. It is recommended to use a tibble
as this prevents common errors such as providing a source by a factor
(\code{data.frame} automatically coerces \code{character} to \code{factor}).}

\item{spec}{Your specification of an indicator. The schemes "<source>-<name>"
and "<source>__<name>" are standard for character but you may enhance
character specification or create alternative class specifications. The
specifications are overwritten if a name or source is provided.}
}
\description{
Indicators are promises of data. They consist of a source and a name that,
together should be unique. The promised data may depend on further attributes
such as the file they may be found in, some kind of selection criterion etc.
}
\section{Workflow}{

If functions like \code{\link[=create]{create()}} encounter a non-indicator, they try to turn it
to an indicator with \code{as_indicator()}. This allows a certain flexibility in
the syntax.
}

\section{Flexibility}{

\code{as_indicator()} and, by extension, \code{\link[=create]{create()}} etc. are intended to be
flexible in syntax. The functions are therefore generic and may be extended
by you. The standard method to specify an indicator is either by explicit
source and name, by the scheme "<source>-<name>" or by the scheme
"<source>__<name>".
}

