% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{is_variable}
\alias{is_variable}
\alias{is_variable_source}
\alias{is_variable_dependent}
\title{Variables}
\usage{
is_variable(x)

is_variable_source(x)

is_variable_dependent(x)
}
\arguments{
\item{x}{an object in R.}
}
\description{
Variables are the fundamental unit of information.
}
\section{Functions}{
\itemize{
\item \code{is_variable}: Is x a variable?

\item \code{is_variable_source}: Is x a source variable?

\item \code{is_variable_dependent}: Is x a dependent variable?
}}

