% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{is_variable}
\alias{is_variable}
\alias{variable}
\title{Variables}
\usage{
is_variable(x)

variable(ind_name = character(0), col = character(0), df_name = "",
  df_key = new_key(), is_key = logical(0), nr_lst = integer(0), ...)
}
\arguments{
\item{x}{an object.}

\item{ind_name}{Names of the indicators}

\item{col}{The columns that belong to the respective indicators.}

\item{df_name}{If a specific column belongs to a data frame with a name, this
column provides the name.}

\item{df_key}{The key of a specific column/variable.}

\item{nr_lst}{integer Internal: gives the list number in which a particular
column can be found.}

\item{...}{other column with variable specific information. If the
information belongs to the indicators, put it in the indicators
collection_df.}

\item{key}{logical. Is a specific variable a key?}
}
\description{
Variables with respect to collections correspond to column names with
respect to data frames. They contain information on every column in every
collection element.
}
