% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-class.R
\name{Structure}
\alias{Structure}
\alias{Structure.default}
\alias{Structure.variable}
\title{Structure}
\usage{
Structure(variables, ...)

\method{Structure}{default}(variables, ...)

\method{Structure}{variable}(variables, relations, edges)
}
\arguments{
\item{variables}{The variables of the structure.}

\item{relations}{relations that extend the structure}

\item{edges}{new connection between contents and relations}
}
\description{
Structures are collections of variables and relations between them -
essentially graphs.
}
