% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-class.R
\name{Structure}
\alias{Structure}
\alias{Structure.default}
\alias{Structure.variable}
\title{Structure}
\usage{
Structure(variables = Content(character(0)), ...)

\method{Structure}{default}(variables, ...)

\method{Structure}{variable}(variables, relations = C_Relation(character(0)),
  edges = data.frame(from = numeric(0), to = numeric(0), name = character(0)))
}
\arguments{
\item{variables}{The variables of the structure.}

\item{relations}{relations that extend the structure}

\item{edges}{new connection between contents and relations}
}
\description{
Structures are collections of variables and relations between them -
essentially graphs.
}
