% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-class.R
\name{Structure}
\alias{Structure}
\alias{Structure.default}
\alias{Structure.relation}
\title{Structure}
\usage{
Structure(relation, ...)

\method{Structure}{default}(relation, ...)

\method{Structure}{relation}(variables)
}
\arguments{
\item{relation}{The relation.}
}
\description{
Structures are collections of variables and relations between them. Saved in
the form \code{relation} \code{content}.
}
