% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-class.R
\name{Structure}
\alias{Structure}
\alias{Structure.default}
\alias{Structure.variable}
\alias{Structure.list}
\title{Structure}
\usage{
Structure(subject, ...)

\method{Structure}{default}(subject, ...)

\method{Structure}{variable}(subject, relation, object, ...)

\method{Structure}{list}(subject, relation, object, ...)
}
\arguments{
\item{relation}{The relation.}

\item{object}{The right variable.}

\item{relation}{The relation.}
}
\description{
Structures are collections of variables and relations between them. Saved in
the form \code{subject} \code{relation} \code{object}.
}
