% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-funs.R, R/shiny-tectr-server.R,
%   R/shiny-tectr-ui.R
\name{tectr}
\alias{tectr}
\alias{tectr_server}
\alias{tectr_server.NULL}
\alias{tectr_ui}
\title{tectr}
\usage{
tectr(project)

tectr_server(project = NULL)

\method{tectr_server}{NULL}(project)

tectr_ui()
}
\arguments{
\item{project}{can be a \code{structure}, a file where \code{tectr()} previously saved
the progress}
}
\description{
The \code{tectr} shiny app provides the main workflow of the package.
}
