% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{variable}
\alias{variable}
\title{Variables}
\usage{
variable(id = character(0), ...)
}
\description{
A variable is a data frame of indicator variables, i. e. every column is
associated with an indicator and every row is associated with an id.
}
